Fix Code Health smells in test/special.js (9.09 -> 10.0)
Split test/special.js (183 functions, "Number of Functions in a Single
Module" smell) into a test/special/ directory, one file per
src/special/ module, mirroring the existing test/mc/ and test/process/
per-module layout. hurwitz-zeta.js and riemann-zeta.js share a single
zeta.js test file since one existing test asserts an identity across
both.

Also extracted a checkReferenceValues helper in error.js to eliminate
a Code Duplication smell between the erf/erfc reference-value tests,
surfaced once the file was small enough for CodeScene to flag it.

No behavior change - same test cases, same assertions. All 13 new
files score 10.0 with no remaining smells.
