--- test/special.js
@@ -1,1805 +1,0 @@
-import { assert } from 'chai'
-import { describe, it } from 'mocha'
-import { equal } from './test-utils.js'
-import * as special from '../src/special/index.js'
-
-describe('special', () => {
-  const EM = 0.5772156649015329
-
-  describe('digamma(z)', () => {
-    it('should return reference values', () => {
-      assert(equal(special.digamma(1), -EM))
-      assert(equal(special.digamma(2), 1 - EM))
-      assert(equal(special.digamma(0.5), -EM - 2 * Math.log(2)))
-      assert(equal(special.digamma(0.25), -Math.PI / 2 - 3 * Math.log(2) - EM))
-      // z >= 10: direct Stirling path
-      assert(equal(special.digamma(10), 2.251752589066721))
-    })
-
-    it('should return reference values for negative non-integer arguments', () => {
-      // Exercises the reflection formula: ψ(z) = ψ(1-z) - π·cot(πz)
-      // ψ(-0.5) = ψ(1.5) = ψ(0.5) + 2 = 2 - EM - 2·ln2
-      assert(equal(special.digamma(-0.5), 2 - EM - 2 * Math.log(2)))
-      // ψ(-1.5) = ψ(2.5) = ψ(1.5) + 2/3
-      assert(equal(special.digamma(-1.5), 2 - EM - 2 * Math.log(2) + 2 / 3))
-      // mpmath mp.dps=50: mp.digamma(-2.5)
-      assert(equal(special.digamma(-2.5), 1.103156640645243, 14))
-    })
-
-    it('should return Infinity at the non-positive integer poles', () => {
-      // ADR-0015: divergence returns Infinity specifically (not NaN, not a huge finite).
-      assert.strictEqual(special.digamma(0), Infinity)
-      assert.strictEqual(special.digamma(-1), Infinity)
-      assert.strictEqual(special.digamma(-2), Infinity)
-    })
-
-    it('should stay full-precision within 1e-6 of a negative integer pole', () => {
-      // mpmath mp.dps=60 evaluated at the exact double of (-1+1e-7) / (-2+1e-7): the input
-      // itself only pins the offset to ~1e-9, but the dominant pole term -1/(z+n) is otherwise
-      // carried at machine precision (no argument-reduction loss in cot(πz)).
-      assert(equal(special.digamma(-1 + 1e-7), -9999999.582479, 13))
-      assert(equal(special.digamma(-2 + 1e-7), -9999999.071376376, 13))
-    })
-
-    it('should satisfy the recurrence ψ(z+1) = ψ(z) + 1/z', () => {
-      for (const z of [0.1, 0.5, 1, 6, 100]) {
-        assert(equal(special.digamma(z + 1), special.digamma(z) + 1 / z))
-      }
-    })
-  })
-
-  describe('.bessel()', () => {
-    it('In(0) should be equal to 0 for n >= 1', () => {
-      for (const n of [1, 2, 3, 5, 10]) {
-        assert(special.besselI(n, 0) === 0)
-      }
-    })
-
-    it('I1(-x) should be equal to -I1(x)', () => {
-      for (const x of [0.1, 1, 9.9, 10, 10.1, 50]) {
-        assert(equal(special.besselI(1, -x), -special.besselI(1, x)))
-      }
-    })
-
-    it('In(-x) should be equal to -In(x) for odd n >= 3', () => {
-      // Regression for the backward-recurrence sign bug fixed in #255: abs(x) with no sign
-      // correction caused I_n(-x) != -I_n(x) for odd n >= 3
-      for (const x of [0.1, 1, 5, 9.9, 10, 10.1, 50]) {
-        assert(equal(special.besselI(3, -x), -special.besselI(3, x)))
-        assert(equal(special.besselI(5, -x), -special.besselI(5, x)))
-        assert(equal(special.besselI(7, -x), -special.besselI(7, x)))
-      }
-    })
-
-    it('I1(x) should match scipy reference values', () => {
-      // scipy.stats cross-validation; these were the values broken by the old _I1 polynomial
-      assert(equal(special.besselI(1, 2), 1.590636854637329))
-      assert(equal(special.besselI(1, 10), 2670.988303701254))
-    })
-
-    it('I1(x) should match reference values at large arguments', () => {
-      // Independent sanity check: leading asymptotic I_1(x) ~ e^x/sqrt(2*pi*x) agrees
-      // with the Miller result to < 1% at x=50 (asymptotic is accurate to ~0.5% with
-      // one correction term), confirming the normalization is not off by any power of e^x.
-      assert(Math.abs(special.besselI(1, 50) * Math.sqrt(2 * Math.PI * 50) / Math.exp(50) - 1) < 0.01)
-      assert(equal(special.besselI(1, 50), 2.9030785901035533e+20))
-      assert(equal(special.besselI(1, 100), 1.0683693903381671e+42))
-      assert(equal(special.besselI(1, 200), 2.0345815493320935e+85))
-    })
-
-    it('I1(x) sign symmetry should hold at large arguments', () => {
-      // Exercises the x < 0 && n % 2 === 1 sign-correction branch for large |x|.
-      assert(equal(special.besselI(1, -50), -special.besselI(1, 50)))
-    })
-
-    it('I0(x) should match reference values at large arguments', () => {
-      // besselI(0, x) also used _I0(x) directly; this confirms the large-x fix.
-      assert(equal(special.besselI(0, 50), 2.9325537838493486e+20))
-      assert(equal(special.besselI(0, 100), 1.073751707131081e+42))
-      assert(equal(special.besselI(0, 200), 2.0396871734097203e+85))
-    })
-
-    it('I0(x) should be continuous across the |x|=10 routing boundary', () => {
-      // Verifies _I0 (|x|<=10) and _besselIBackward (|x|>10) agree near the crossover.
-      assert(equal(special.besselI(0, 9.9), 2560.9633532560433))
-      assert(equal(special.besselI(0, 10), 2815.716628466255))
-      // mpmath mp.dps=50: besseli(0, mpf(10.1)) -> 3095.9756707889824
-      assert(equal(special.besselI(0, 10.1), 3095.9756707889824))
-    })
-
-    it('I0(x) should match mpmath reference values in the Miller warm-up band (10, 15]', () => {
-      // Issue #1185: _besselIBackward's j_max margin term degenerates to 0 at n=0
-      // (sqrt(40*0)=0), leaving a ~1e-9 relative-error precision gap in this band.
-      // mpmath mp.dps=50: besseli(0, mpf(x))
-      assert(equal(special.besselI(0, 10.5), 4527.441714638888, 14))
-      assert(equal(special.besselI(0, 11), 7288.489339821248, 14))
-      assert(equal(special.besselI(0, 12), 18948.925349296307, 14))
-      assert(equal(special.besselI(0, 13), 49444.489582217575, 14))
-      assert(equal(special.besselI(0, 14), 129418.56270064856, 14))
-    })
-  })
-
-  describe('.besselISpherical()', () => {
-    it('i(0, 0) should be 1', () => {
-      assert(special.besselISpherical(0, 0) === 1)
-    })
-
-    it('i(1, 0) should be 0', () => {
-      assert(special.besselISpherical(1, 0) === 0)
-    })
-
-    it('i(n, 0) should be 0 for n > 0', () => {
-      for (const n of [1, 2, 5]) {
-        assert(special.besselISpherical(n, 0) === 0)
-      }
-    })
-
-    it('should satisfy the recurrence relation for negative order', () => {
-      for (const n of [-1, -2, -3, -5]) {
-        for (const x of [0.5, 2, 9, 11]) {
-          assert(equal(special.besselISpherical(n - 1, x) - special.besselISpherical(n + 1, x),
-            (2 * n + 1) * special.besselISpherical(n, x) / x))
-        }
-      }
-    })
-
-    it('should satisfy the recurrence relation for positive order', () => {
-      for (const n of [1, 2, 3, 5]) {
-        for (const x of [0.5, 2, 9, 11]) {
-          assert(equal(special.besselISpherical(n - 1, x) - special.besselISpherical(n + 1, x),
-            (2 * n + 1) * special.besselISpherical(n, x) / x))
-        }
-      }
-    })
-
-    it('should return accurate small-x values for n=1', () => {
-      // i_1(x) = x/3 + x³/30 + x⁵/840 + ...; (2*1+1)!! = 3
-      assert(special.besselISpherical(1, 0) === 0)
-      // At x=1e-6 the second term is ~3.7e-20, negligible vs 1e-10 relative tolerance
-      assert(equal(special.besselISpherical(1, 1e-6), 1e-6 / 3, 10))
-      // At x=1e-3 include two terms (third term ~3.6e-15 relative)
-      assert(equal(special.besselISpherical(1, 1e-3), 1e-3 / 3 + 1e-9 / 30, 10))
-      // At x=0.1 (Taylor branch): hand-computed from series
-      assert(equal(special.besselISpherical(1, 0.1), 0.03336667857363341, 10))
-    })
-
-    it('should return accurate small-x values for n=2', () => {
-      // i_2(x) = x²/15 + x⁴/210 + ...; (2*2+1)!! = 5!! = 15
-      assert(special.besselISpherical(2, 0) === 0)
-      // At x=1e-6 the second term is ~7e-14 relative
-      assert(equal(special.besselISpherical(2, 1e-6), 1e-12 / 15, 10))
-      // At x=1e-3 include two terms x²/15 + x⁴/210 (third term ~2e-15 relative)
-      assert(equal(special.besselISpherical(2, 1e-3), 1e-6 / 15 + 1e-12 / 210, 10))
-      // At x=0.1 (Taylor branch): hand-computed from series
-      assert(equal(special.besselISpherical(2, 0.1), 6.671429894380334e-4, 10))
-    })
-
-    it('should return accurate small-x values for n=5', () => {
-      // i_5(x) = x⁵/10395 + ...; (2*5+1)!! = 11!! = 10395, 13!! = 135135
-      assert(special.besselISpherical(5, 0) === 0)
-      // At x=1e-6 the second term is ~4e-14 relative
-      assert(equal(special.besselISpherical(5, 1e-6), 1e-30 / 10395, 10))
-      // At x=1e-3 include two terms x⁵/10395 + x⁷/270270 (third term ~6e-16 relative)
-      assert(equal(special.besselISpherical(5, 1e-3), 1e-15 / 10395 + 1e-21 / 270270, 10))
-      // At x=0.1 (Taylor branch): hand-computed from series
-      assert(equal(special.besselISpherical(5, 0.1), 9.62371024043737e-10, 10))
-    })
-
-    it('should satisfy positive-order recurrence in the Taylor branch', () => {
-      for (const n of [1, 2, 5]) {
-        for (const x of [0.01, 0.1, 0.5, 0.99]) {
-          assert(equal(special.besselISpherical(n - 1, x) - special.besselISpherical(n + 1, x),
-            (2 * n + 1) * special.besselISpherical(n, x) / x))
-        }
-      }
-    })
-
-    it('should return accurate large-argument values', () => {
-      // i_n(x) = sqrt(pi/(2x)) * I_{n+1/2}(x); mpmath mp.dps=50:
-      // i_3(50) = 4.5930269336647156e+19, i_3(100) = 1.2654109836313822e+41
-      // i_7(50) = 2.947924492475899e+19, i_7(100) = 1.0145184556029379e+41
-      assert(equal(special.besselISpherical(3, 50), 4.5930269336647156e+19, 10))
-      assert(equal(special.besselISpherical(3, 100), 1.2654109836313822e+41, 10))
-      assert(equal(special.besselISpherical(7, 50), 2.947924492475899e+19, 10))
-      assert(equal(special.besselISpherical(7, 100), 1.0145184556029379e+41, 10))
-    })
-
-    it('should return accurate values past _hi\'s pre-#1311 fixed MAX_ITER=100 budget (#1311)', () => {
-      // 7*sqrt(x)+20 only exceeds MAX_ITER=100 once x > ~131 (e.g. it's still only 90 at
-      // x=100), so x=300 and x=500 are chosen to land past that crossover -- and past the
-      // x~250 point where the pre-#1311 fixed budget silently under-converged (CHANGELOG) --
-      // unlike x=50/100 above, which pass under the old, un-widened budget too.
-      // i_n(x) = sqrt(pi/(2x)) * I_{n+1/2}(x); mpmath mp.dps=50, rounded to float64:
-      // i_3(300) = 3.1731675432386596e+127, i_3(500) = 1.3868331583406083e+214
-      // i_7(300) = 2.9484482410057756e+127, i_7(500) = 1.3270783008491142e+214
-      assert(equal(special.besselISpherical(3, 300), 3.1731675432386596e+127, 10))
-      assert(equal(special.besselISpherical(3, 500), 1.3868331583406083e+214, 10))
-      assert(equal(special.besselISpherical(7, 300), 2.9484482410057756e+127, 10))
-      assert(equal(special.besselISpherical(7, 500), 1.3270783008491142e+214, 10))
-    })
-
-    // _assertHiConverged's throw was probed empirically across n in [1, 1e6] and x in
-    // [1e-10, 709] (the exp(x) overflow ceiling for the unscaled function) through both
-    // besselISpherical and besselISphericalExpScaled -- including large n relative to x, per
-    // the "depth grows with x, not n" claim in _hi's comment -- and never fired: 7*sqrt(x)+20
-    // keeps a comfortable margin over actual convergence depth throughout the whole domain
-    // besselISpherical's Wronskian branch is reached from. The check remains defensive: it
-    // guards the 7*sqrt(x)+20 margin against being falsified by a future change, not a path
-    // this test suite can currently force through valid input.
-
-    it('should satisfy the parity identity i_n(-x) = (-1)^n * i_n(x) for n > 1 (#1324)', () => {
-      // i_n(x) is an entire function of x -- its Taylor series has only x^(n+2k) terms --
-      // so it has definite parity (-1)^n. In the Taylor branch (|x| < 1), negative x is
-      // computed through the *same* series (no separate code path), so cross-checking it
-      // against the positive-x call is a genuine consistency test, not a tautology.
-      for (const n of [2, 3, 5, 7]) {
-        for (const x of [0.1, 0.5]) {
-          assert(equal(special.besselISpherical(n, -x), (n % 2 === 0 ? 1 : -1) * special.besselISpherical(n, x)))
-        }
-      }
-      // mpmath mp.dps=50: i_n(x) = sqrt(pi/(2x)) * I_{n+1/2}(x); i_2(0.5) = 0.01696636036086198
-      assert(equal(special.besselISpherical(2, -0.5), 0.01696636036086198))
-
-      // The Wronskian/_hi branch (|x| >= 1) is where #1324's bug actually lived: _hi(n, x)
-      // computed Math.sqrt(x) directly on the unmapped negative x, turning its iteration
-      // budget into NaN. besselISpherical(n, -x) maps internally to (-1)^n * besselISpherical(n, x),
-      // so asserting against that same formula would be tautological -- it would pass even if
-      // the underlying _hi/_kn computation were wrong. These are independently sourced from
-      // mpmath mp.dps=50 via i_n(x) = sqrt(pi/(2x)) * I_{n+1/2}(x), never derived from ranjs:
-      // i_2(2) = 0.3518560885534178, i_3(2) = 0.09474252219651647, i_7(2) = 7.097944523040642e-05
-      // i_2(9) = 316.7872223410896, i_5(9) = 81.7183759470097
-      assert(equal(special.besselISpherical(2, -2), 0.3518560885534178))
-      assert(equal(special.besselISpherical(3, -2), -0.09474252219651647))
-      assert(equal(special.besselISpherical(7, -2), -7.097944523040642e-05))
-      assert(equal(special.besselISpherical(2, -9), 316.7872223410896))
-      assert(equal(special.besselISpherical(5, -9), -81.7183759470097))
-      // mpmath mp.dps=50, negated per the parity identity (n odd, matching the "should return
-      // accurate large-argument values" test's i_3(50)/i_7(50) above): i_3(50) = 4.5930269336647156e+19
-      // -> i_3(-50) = -4.5930269336647156e+19; i_7(50) = 2.947924492475899e+19 -> i_7(-50) = -2.947924492475899e+19
-      assert(equal(special.besselISpherical(3, -50), -4.5930269336647156e+19, 10))
-      assert(equal(special.besselISpherical(7, -50), -2.947924492475899e+19, 10))
-    })
-  })
-
-  // Issue #1310: besselIExpScaled (added by #1292) was previously only reachable indirectly
-  // through NoncentralChi2/NoncentralChi pdf evaluation at a handful of large-lambda*x points,
-  // leaving its own branch structure (n=0 small-x/large-x split, x=0, negative-x sign flip)
-  // without any dedicated coverage.
-  describe('.besselIExpScaled()', () => {
-    it('should return exactly 0 at x=0 for n >= 1', () => {
-      for (const n of [1, 2, 3, 5]) {
-        assert(special.besselIExpScaled(n, 0) === 0)
-      }
-    })
-
-    it('should return 1 at n=0, x=0', () => {
-      // _I0(0) * exp(-0) = 1 * 1 = 1
-      assert(special.besselIExpScaled(0, 0) === 1)
-    })
-
-    it('should return accurate values in the n=0 small-x _I0 branch (|x| <= 10)', () => {
-      // mpmath mp.dps=50: besseli(0, mpf(x)) * exp(-mpf(x)). Deliberately not checked via
-      // besselI(0, x) * Math.exp(-x): both besselI and besselIExpScaled call the exact same
-      // _I0(x) helper in this branch, so that identity would hold even if _I0 itself were wrong.
-      assert(equal(special.besselIExpScaled(0, 0.001), 0.9990007495835156, 13))
-      assert(equal(special.besselIExpScaled(0, 0.5), 0.6450352704491501, 13))
-      assert(equal(special.besselIExpScaled(0, 1), 0.46575960759364043, 13))
-      assert(equal(special.besselIExpScaled(0, 5), 0.18354081260932836, 13))
-      assert(equal(special.besselIExpScaled(0, 9.9), 0.1284955220071385, 13))
-      assert(equal(special.besselIExpScaled(0, 10), 0.1278333371634286, 13))
-    })
-
-    it('should return accurate values in the n=0 large-x _besselIBackward branch (|x| > 10)', () => {
-      // mpmath mp.dps=50: besseli(0, mpf(x)) * exp(-mpf(x)). Not checked via besselI(0, x) *
-      // Math.exp(-x): both share _besselIBackward's Miller-recurrence loop, differing only in
-      // the final scaled ? y/sum : y*exp(x-log(sum)) line, so a bug in the shared loop would
-      // still satisfy that identity.
-      assert(equal(special.besselIExpScaled(0, 10.1), 0.12718130354436347, 13))
-      assert(equal(special.besselIExpScaled(0, 14), 0.10761525167069509, 13))
-      assert(equal(special.besselIExpScaled(0, 50), 0.05656162664745419, 13))
-      assert(equal(special.besselIExpScaled(0, 200), 0.028227159949111916, 13))
-      assert(equal(special.besselIExpScaled(0, 700), 0.015081295651531358, 13))
-    })
-
-    it('should be continuous across the |x|=10 n=0 routing boundary', () => {
-      // mpmath mp.dps=50: besseli(0, mpf(x)) * exp(-mpf(x))
-      assert(equal(special.besselIExpScaled(0, 9.9), 0.1284955220071385, 13))
-      assert(equal(special.besselIExpScaled(0, 10), 0.1278333371634286, 13))
-      assert(equal(special.besselIExpScaled(0, 10.1), 0.12718130354436347, 13))
-    })
-
-    it('should return accurate values for n != 0 across the _besselIBackward crossover', () => {
-      // mpmath mp.dps=50: besseli(n, mpf(x)) * exp(-mpf(x)). Not checked via besselI(n, x) *
-      // Math.exp(-x): both share the same _besselIBackward Miller-recurrence loop for any given
-      // (n, x), so that identity cannot catch a bug in the shared loop itself (only in the
-      // scaled=true/false post-processing, which is a single line).
-      assert(equal(special.besselIExpScaled(1, 0.5), 0.1564208031848717, 13))
-      assert(equal(special.besselIExpScaled(1, 9.9), 0.12182203639796144, 13))
-      assert(equal(special.besselIExpScaled(1, 10.1), 0.12071085823848707, 13))
-      assert(equal(special.besselIExpScaled(1, 100), 0.03974415302513025, 13))
-      assert(equal(special.besselIExpScaled(1, 500), 0.017827851852898056, 13))
-      assert(equal(special.besselIExpScaled(2, 0.5), 0.01935205770966328, 13))
-      assert(equal(special.besselIExpScaled(2, 9.9), 0.10388500960350994, 13))
-      assert(equal(special.besselIExpScaled(2, 10.1), 0.1032781632991185, 13))
-      assert(equal(special.besselIExpScaled(5, 0.5), 4.987605521470164e-06, 13))
-      assert(equal(special.besselIExpScaled(5, 9.9), 0.03500619134149665, 13))
-      assert(equal(special.besselIExpScaled(5, 10.1), 0.03555744332660292, 13))
-      assert(equal(special.besselIExpScaled(10, 0.5), 1.6030859629529217e-13, 10))
-      assert(equal(special.besselIExpScaled(10, 9.9), 0.0009555334260008234, 13))
-      assert(equal(special.besselIExpScaled(10, 10.1), 0.0010330372238920365, 13))
-      assert(equal(special.besselIExpScaled(10, 500), 0.016145898955259176, 13))
-    })
-
-    it('should stay finite well past besselI\'s own ~709 overflow ceiling', () => {
-      // besselI(0, 1000) overflows to Infinity; the whole reason besselIExpScaled exists (#1292)
-      // is to stay representable here. mpmath mp.dps=50: besseli(n, mpf(x)) * exp(-mpf(x))
-      assert.strictEqual(special.besselI(0, 1000), Infinity)
-      assert(equal(special.besselIExpScaled(0, 1000), 0.012617240455891257, 13))
-      assert(equal(special.besselIExpScaled(3, 1000), 0.01256056218254712, 13))
-      assert(equal(special.besselIExpScaled(0, 5000), 0.005642036898744589, 13))
-      assert(equal(special.besselIExpScaled(3, 5000), 0.0056369608425911434, 13))
-    })
-
-    it('should satisfy sign symmetry for odd n at negative x', () => {
-      for (const n of [1, 3, 5]) {
-        for (const x of [0.1, 1, 5, 9.9, 10, 10.1, 50, 500]) {
-          assert(equal(special.besselIExpScaled(n, -x), -special.besselIExpScaled(n, x)))
-        }
-      }
-    })
-  })
-
-  // Issue #1321: besselIExpScaled(n, x) itself underflows to exactly 0 once the Bessel order n
-  // is large relative to the argument x (e.g. n=999, x~63.25), because the true scaled value is
-  // genuinely non-representable as a double there -- yet its logarithm is a normal finite
-  // number. logBesselIExpScaled fills that gap: log(exp(-x) * I_n(x)), computed directly in
-  // log-space so callers (Skellam._pdf) can combine it with other log-scale terms before a
-  // single final exponentiation.
-  describe('.logBesselIExpScaled()', () => {
-    it('should return 0 at n=0, x=0', () => {
-      assert(special.logBesselIExpScaled(0, 0) === 0)
-    })
-
-    it('should return -Infinity at x=0 for n >= 1', () => {
-      for (const n of [1, 2, 3, 5]) {
-        assert(special.logBesselIExpScaled(n, 0) === -Infinity)
-      }
-    })
-
-    it('should agree with log(besselIExpScaled(n, x)) in the direct-delegation regime', () => {
-      // mpmath mp.dps=50: log(besseli(n, mpf(x))) - x
-      assert(equal(special.logBesselIExpScaled(0, 0.5), -0.4384502808145187, 12))
-      assert(equal(special.logBesselIExpScaled(1, 0.5), -1.8552054470253345, 12))
-      assert(equal(special.logBesselIExpScaled(5, 5), -4.230829927801479, 12))
-      assert(equal(special.logBesselIExpScaled(10, 100), -3.722366634346062, 12))
-      // Sanity: this regime is the one where besselIExpScaled already returns a nonzero value.
-      assert(special.besselIExpScaled(10, 100) > 0)
-    })
-
-    it('should return a finite value in the large-order/small-argument regime where besselIExpScaled underflows to exactly 0', () => {
-      // twoSqrtProd for Skellam(1000,1)/(2000,1)/(5000,1) -- the exact (n, x) pairs issue #1321
-      // reports as NaN. Confirms the underflow precondition, then checks the finite log value.
-      const z1 = 2 * Math.sqrt(1000)
-      const z2 = 2 * Math.sqrt(2000)
-      const z3 = 2 * Math.sqrt(5000)
-      assert.strictEqual(special.besselIExpScaled(999, z1), 0)
-      assert.strictEqual(special.besselIExpScaled(990, z1), 0)
-      assert.strictEqual(special.besselIExpScaled(1999, z2), 0)
-      assert.strictEqual(special.besselIExpScaled(4999, z3), 0)
-
-      // mpmath mp.dps=50: log(besseli(n, mpf(x))) - x
-      assert(equal(special.logBesselIExpScaled(999, z1), -2517.0427133980397, 12))
-      assert(equal(special.logBesselIExpScaled(990, z1), -2485.9938897835273, 12))
-      assert(equal(special.logBesselIExpScaled(1999, z2), -5690.264408550746, 12))
-      assert(equal(special.logBesselIExpScaled(4999, z3), -16434.323389931138, 12))
-      assert(equal(special.logBesselIExpScaled(100, 20), -152.4955121081758, 12))
-    })
-  })
-
-  // Issue #1310: besselISphericalExpScaled (added by #1292) had the same coverage gap as
-  // besselIExpScaled above -- only reachable indirectly through NoncentralChi/NoncentralChi2.
-  describe('.besselISphericalExpScaled()', () => {
-    it('should return 1 at n=0, x=0', () => {
-      assert(special.besselISphericalExpScaled(0, 0) === 1)
-    })
-
-    it('should return exactly 0 at x=0 for n >= 1', () => {
-      for (const n of [1, 2, 3, 5]) {
-        assert(special.besselISphericalExpScaled(n, 0) === 0)
-      }
-    })
-
-    it('should equal exp(-x) * besselISpherical(n, x) in the n=0 closed-form branch', () => {
-      // (1 - exp(-2x)) / (2x) == exp(-x) * sinh(x) / x, checked against the unscaled function's
-      // own independent code path.
-      for (const x of [0.001, 0.5, 1, 5, 10, 50, 100, 500]) {
-        assert(equal(special.besselISphericalExpScaled(0, x), special.besselISpherical(0, x) * Math.exp(-x)))
-      }
-    })
-
-    it('should equal exp(-x) * besselISpherical(n, x) in the n=1 closed-form branch (|x| >= 1)', () => {
-      // Legitimate cross-check: for |x| >= 1, besselISpherical(1, x) uses (cosh(x)-sinh(x)/x)/x
-      // while besselISphericalExpScaled(1, x) uses the algebraically-equivalent but separately
-      // coded (1+e^-2x)/(2x) - (1-e^-2x)/(2x^2) -- these are genuinely independent formulas,
-      // unlike the |x| < 1 Taylor branch below (which shares _besselISphericalTaylor exactly).
-      for (const x of [1, 1.01, 1.1, 2, 5, 50, 500]) {
-        assert(equal(special.besselISphericalExpScaled(1, x), special.besselISpherical(1, x) * Math.exp(-x)))
-      }
-    })
-
-    it('should return accurate small-x Taylor values for n=1 (|x| < 1)', () => {
-      // mpmath mp.dps=50: sqrt(pi/(2x)) * besseli(1.5, x) * exp(-x). Not checked via
-      // besselISpherical(1, x) * Math.exp(-x): both share the exact same
-      // _besselISphericalTaylor(1, x) helper call in this branch.
-      assert(equal(special.besselISphericalExpScaled(1, 1e-6), 3.3333300000019996e-07, 10))
-      assert(equal(special.besselISphericalExpScaled(1, 1e-3), 0.00033300019991114283, 10))
-      assert(equal(special.besselISphericalExpScaled(1, 0.1), 0.03019141928900223, 10))
-      assert(equal(special.besselISphericalExpScaled(1, 0.5), 0.10363832351432696, 10))
-      assert(equal(special.besselISphericalExpScaled(1, 0.9), 0.13214067137099655, 10))
-      assert(equal(special.besselISphericalExpScaled(1, 0.99), 0.13506671882903618, 10))
-    })
-
-    it('should be continuous across the n=1 |x|=1 Taylor/closed-form crossover', () => {
-      // mpmath mp.dps=50: sqrt(pi/(2x)) * besseli(1.5, x) * exp(-x)
-      assert(equal(special.besselISphericalExpScaled(1, 0.99), 0.13506671882903618, 13))
-      assert(equal(special.besselISphericalExpScaled(1, 1), 0.1353352832366127, 13))
-      assert(equal(special.besselISphericalExpScaled(1, 1.01), 0.13559331673906708, 13))
-    })
-
-    it('should return accurate values in the n>=2 small-x Taylor branch (|x| < 1)', () => {
-      // mpmath mp.dps=50: sqrt(pi/(2x)) * besseli(n+0.5, x) * exp(-x). Not checked via
-      // besselISpherical(n, x) * Math.exp(-x): both share the exact same
-      // _besselISphericalTaylor(n, x) helper call in this branch.
-      assert(equal(special.besselISphericalExpScaled(2, 0.001), 6.660003807937037e-08, 10))
-      assert(equal(special.besselISphericalExpScaled(2, 0.5), 0.01029061774259589, 10))
-      assert(equal(special.besselISphericalExpScaled(2, 0.99), 0.026025479653984947, 10))
-      assert(equal(special.besselISphericalExpScaled(3, 0.001), 9.514291003175277e-12, 10))
-      assert(equal(special.besselISphericalExpScaled(3, 0.5), 0.0007321460883680679, 10))
-      assert(equal(special.besselISphericalExpScaled(5, 0.001), 9.610394788422048e-20, 10))
-      assert(equal(special.besselISphericalExpScaled(5, 0.5), 1.8409903951734993e-06, 10))
-    })
-
-    it('should return accurate values in the n>=2 Wronskian branch (|x| >= 1)', () => {
-      // mpmath mp.dps=50: sqrt(pi/(2x)) * besseli(n+0.5, x) * exp(-x). Not checked via
-      // besselISpherical(n, x) * Math.exp(-x): both derive from the same _hi(n+1, x) and
-      // _knRaw(n+1, x) calls, differing only in whether the exp(-x)/x normalization _kn applies
-      // is included -- a bug in the shared _hi/_knRaw recurrences would still satisfy that
-      // identity.
-      assert(equal(special.besselISphericalExpScaled(2, 1.01), 0.026626056675978112, 10))
-      assert(equal(special.besselISphericalExpScaled(2, 2), 0.04761854340290348, 10))
-      assert(equal(special.besselISphericalExpScaled(2, 10), 0.036499999862933286, 10))
-      assert(equal(special.besselISphericalExpScaled(2, 500), 0.000994012, 13))
-      assert(equal(special.besselISphericalExpScaled(3, 1.01), 0.0037811549767991933, 10))
-      assert(equal(special.besselISphericalExpScaled(3, 10), 0.026750000181896806, 10))
-      assert(equal(special.besselISphericalExpScaled(3, 500), 0.00098805988, 13))
-      assert(equal(special.besselISphericalExpScaled(5, 1.01), 3.829481966569938e-05, 10))
-      assert(equal(special.besselISphericalExpScaled(5, 10), 0.01075250041985184, 10))
-      assert(equal(special.besselISphericalExpScaled(5, 500), 0.00097041665508976, 13))
-    })
-
-    it('should stay finite well past besselISpherical\'s own overflow ceiling', () => {
-      // mpmath mp.dps=50: sqrt(pi/(2x)) * besseli(3.5, x) * exp(-x)
-      assert(equal(special.besselISphericalExpScaled(3, 700), 0.0007081850999583507, 12))
-      assert(equal(special.besselISphericalExpScaled(3, 5000), 9.9880059988e-05, 12))
-    })
-
-    it('should return accurate values in the n<0 backward-recurrence branch', () => {
-      // mpmath mp.dps=50: sqrt(pi/(2x)) * besseli(n+0.5, x) * exp(-x). besselISphericalExpScaled
-      // has no independent closed form for n < 0 -- the production code's own recursion IS the
-      // implementation, and a recurrence-relation identity check against that same recursion
-      // (as besselISpherical's own negative-order test above does) would be tautological here:
-      // it's an algebraic rearrangement of the exact recursion being tested, so it can't catch a
-      // bug in the recursion's own coefficients. These literals are the only independent check.
-      assert(equal(special.besselISphericalExpScaled(-1, 0.5), 1.3678794411714423, 13))
-      assert(equal(special.besselISphericalExpScaled(-1, 5), 0.10000453999297625, 13))
-      assert(equal(special.besselISphericalExpScaled(-1, 50), 0.01, 13))
-      assert(equal(special.besselISphericalExpScaled(-2, 0.5), -2.103638323514327, 13))
-      assert(equal(special.besselISphericalExpScaled(-2, 5), 0.0799945520084285, 13))
-      assert(equal(special.besselISphericalExpScaled(-2, 50), 0.0098, 13))
-      assert(equal(special.besselISphericalExpScaled(-3, 0.5), 13.989709382257404, 13))
-      assert(equal(special.besselISphericalExpScaled(-3, 5), 0.05200780878791915, 13))
-      assert(equal(special.besselISphericalExpScaled(-3, 50), 0.009412, 13))
-    })
-  })
-
-  describe('.besselInu()', () => {
-    it('should return accurate small-x values (regression)', () => {
-      // These values must not change — they confirm the Taylor series is undisturbed at small x.
-      assert(equal(special.besselInu(0.5, 1), 0.9376748882454871))
-      assert(equal(special.besselInu(1.5, 1), 0.29352532634747946))
-      assert(equal(special.besselInu(2.3, 1), 0.08157483645893206))
-      assert(equal(special.besselInu(0.5, 10), 2778.7846038745683))
-      assert(equal(special.besselInu(1.5, 10), 2500.906154942116))
-      assert(equal(special.besselInu(2.3, 10), 2132.6900841622582))
-    })
-
-    it('should match the asymptotic leading term for nu=0.5 at x=50', () => {
-      // I_{0.5}(x) ~ e^x / sqrt(2*pi*x) exactly (all correction terms vanish for nu=0.5
-      // because mu = 4*0.25 = 1, so mu-1 = 0). Relative error < 1% verifies the
-      // normalization is not off by any power of e^x.
-      assert(Math.abs(special.besselInu(0.5, 50) * Math.sqrt(2 * Math.PI * 50) / Math.exp(50) - 1) < 0.01)
-    })
-
-    it('should return accurate large-argument values for nu=0.5 (exact formula)', () => {
-      // I_{0.5}(x) = sqrt(2/(pi*x)) * sinh(x); these are cross-validated against that formula.
-      assert(equal(special.besselInu(0.5, 50), 2.9251568529912876e+20))
-      assert(equal(special.besselInu(0.5, 100), 1.0724035825423096e+42))
-      assert(equal(special.besselInu(0.5, 200), 2.0384095654829366e+85))
-    })
-
-    it('should return accurate large-argument values for nu=1.5 (exact formula)', () => {
-      // I_{1.5}(x) = sqrt(2/(pi*x)) * (cosh(x) - sinh(x)/x); cross-validated against that formula.
-      assert(equal(special.besselInu(1.5, 50), 2.866653715931459e+20))
-      assert(equal(special.besselInu(1.5, 100), 1.0616795467168857e+42))
-      assert(equal(special.besselInu(1.5, 200), 2.0282175176555217e+85))
-    })
-
-    it('should return accurate large-argument values for nu=2.3 (asymptotic-validated)', () => {
-      // No closed form; values computed by the Taylor series and verified to 1e-15 relative
-      // error against the DLMF 10.40.1 asymptotic expansion with optimal truncation.
-      assert(equal(special.besselInu(2.3, 50), 2.779977151326617e+20))
-      assert(equal(special.besselInu(2.3, 100), 1.0455847305178129e+42))
-      assert(equal(special.besselInu(2.3, 200), 2.0128232824293037e+85))
-    })
-
-    it('should return finite values for very negative fractional order near the x~710 overflow boundary (issue #1215)', () => {
-      // mpmath mp.dps=50: besseli(nu, x). Previously returned Infinity because the
-      // recursiveSum accumulator overflowed before the tiny (x/2)^nu prefactor was applied.
-      assert(equal(special.besselInu(-1.5, 709), 1.2295937306183464e+306, 13))
-      assert(equal(special.besselInu(-2.5, 700), 1.5227751694938985e+302, 13))
-      assert(equal(special.besselInu(-3.3, 710), 3.3197593551403374e+306, 13))
-    })
-
-    it('should handle the x=0 boundary for all three sign/order sub-cases', () => {
-      // nu > 0: Math.pow(0, nu) = 0, gamma(nu+1) finite and positive => I_nu(0) = 0.
-      assert.strictEqual(special.besselInu(2.3, 0), 0)
-
-      // nu === 0: Math.pow(0, 0) === 1 (JS convention) and gamma(1) = 1 mathematically;
-      // ranjs' gamma() is a floating-point approximation (not exactly 1 at x=1), so this
-      // uses the tolerant `equal` helper rather than strictEqual.
-      assert(equal(special.besselInu(0, 0), 1))
-
-      // nu = -1.5 (non-integer, negative): gamma(-0.5) = -2*sqrt(pi) ~= -3.5449077018110322
-      // (negative), and Math.pow(0, -1.5) is +Infinity per IEEE 754 (+0 raised to a negative
-      // exponent), so the product +Infinity * (1 / negative) diverges to -Infinity.
-      assert.strictEqual(special.besselInu(-1.5, 0), -Infinity)
-    })
-  })
-
-  describe('.besselK()', () => {
-    it('K_n(0) should be Infinity for all n', () => {
-      // K_ν diverges at x=0 for all ν: ADR-0015 divergence → Infinity
-      for (const n of [0, 1, 2, 3]) {
-        assert.strictEqual(special.besselK(n, 0), Infinity)
-      }
-    })
-
-    it('K_0(x) should match mpmath reference values', () => {
-      // mpmath mp.dps=50: besselk(0, x) for x in {0.5, 1, 5, 10}
-      // x=10 uses asymptotic; optimal-truncation error bound ~3.7e-10 limits to 9 sig figs there
-      assert(equal(special.besselK(0, 0.5), 0.9244190712276659))
-      assert(equal(special.besselK(0, 1), 0.42102443824070834))
-      assert(equal(special.besselK(0, 5), 0.0036910983340425942))
-      assert(equal(special.besselK(0, 10), 1.778006231616765e-05, 9))
-    })
-
-    it('K_1(x) should match mpmath reference values', () => {
-      // mpmath mp.dps=50: besselk(1, x)
-      // x=10 uses asymptotic; optimal-truncation error bound limits to 9 sig figs there
-      assert(equal(special.besselK(1, 0.5), 1.656441120003301))
-      assert(equal(special.besselK(1, 1), 0.6019072301972346))
-      assert(equal(special.besselK(1, 5), 0.004044613445452165))
-      assert(equal(special.besselK(1, 10), 1.8648773453825585e-05, 9))
-    })
-
-    it('K_n(x) should match mpmath reference values for n=2,3,4', () => {
-      // mpmath mp.dps=50: besselk(n, x)
-      assert(equal(special.besselK(2, 1), 1.6248388986351774))
-      assert(equal(special.besselK(3, 1), 7.101262824737945))
-      assert(equal(special.besselK(4, 1), 44.232415847062846))
-      assert(equal(special.besselK(2, 5), 0.00530894371222346))
-      assert(equal(special.besselK(3, 5), 0.008291768415230933))
-    })
-
-    it('K_0(x) should satisfy the asymptotic leading term at large x', () => {
-      // K_ν(x) ~ sqrt(π/(2x)) * exp(-x) for large x; 1% tolerance at x=50
-      const x = 50
-      assert(Math.abs(special.besselK(0, x) * Math.exp(x) * Math.sqrt(2 * x / Math.PI) - 1) < 0.01)
-    })
-
-    it('should satisfy the recurrence K_{n+1}(x) = (2n/x)*K_n(x) + K_{n-1}(x)', () => {
-      for (const [n, x] of [[1, 1], [2, 1], [1, 5], [2, 5]]) {
-        const lhs = special.besselK(n + 1, x)
-        const rhs = (2 * n / x) * special.besselK(n, x) + special.besselK(n - 1, x)
-        assert(equal(lhs, rhs))
-      }
-    })
-
-    it('K_0 and K_1 should be continuous across the series/asymptotic crossover at x=6', () => {
-      // x≤6 uses combined series; x>6 uses asymptotic (_X_K_SERIES=6)
-      // Over a 0.02 step the natural variation is ~2.2% (K decreases at ~1/unit near x=6)
-      // A large discontinuity would produce a ratio outside [1, 1.1] or a sign flip
-      const k0b = special.besselK(0, 5.99)
-      const k0a = special.besselK(0, 6.01)
-      const k1b = special.besselK(1, 5.99)
-      const k1a = special.besselK(1, 6.01)
-      assert(k0b > k0a && k0b / k0a < 1.1)
-      assert(k1b > k1a && k1b / k1a < 1.1)
-    })
-  })
-
-  describe('.besselKnu()', () => {
-    it('K_nu(0) should be Infinity', () => {
-      assert.strictEqual(special.besselKnu(0.5, 0), Infinity)
-      assert.strictEqual(special.besselKnu(1.5, 0), Infinity)
-    })
-
-    it('K_{0.5}(x) should match the exact closed form sqrt(pi/(2x))*exp(-x)', () => {
-      // K_{1/2}(x) = sqrt(pi/(2x)) * exp(-x) exactly (DLMF 10.39.2)
-      for (const x of [1, 5, 10, 50]) {
-        const exact = Math.sqrt(Math.PI / (2 * x)) * Math.exp(-x)
-        assert(equal(special.besselKnu(0.5, x), exact))
-      }
-    })
-
-    it('K_{1.5}(x) should match the exact closed form sqrt(pi/(2x))*exp(-x)*(1+1/x)', () => {
-      // K_{3/2}(x) = sqrt(pi/(2x)) * exp(-x) * (1 + 1/x) exactly (DLMF 10.39.2)
-      for (const x of [1, 5, 10, 50]) {
-        const exact = Math.sqrt(Math.PI / (2 * x)) * Math.exp(-x) * (1 + 1 / x)
-        assert(equal(special.besselKnu(1.5, x), exact))
-      }
-    })
-
-    it('K_{2.5}(x) should match mpmath reference values', () => {
-      // mpmath mp.dps=50: besselk(2.5, x)
-      // x=10 uses asymptotic; for nu=2.5 the series terminates at k=3 (4nu²-25=0),
-      // so the only error is floating-point rounding — precision=10 is correct here
-      assert(equal(special.besselKnu(2.5, 1), 3.2274795311352618))
-      assert(equal(special.besselKnu(2.5, 5), 0.006495775004385758))
-      assert(equal(special.besselKnu(2.5, 10), 2.393132586462789e-05))
-    })
-
-    it('should dispatch to besselK for integer nu', () => {
-      for (const x of [1, 5]) {
-        assert.strictEqual(special.besselKnu(0, x), special.besselK(0, x))
-        assert.strictEqual(special.besselKnu(1, x), special.besselK(1, x))
-        assert.strictEqual(special.besselKnu(2, x), special.besselK(2, x))
-      }
-    })
-
-    it('should dispatch negative-integer nu to the correct absolute order', () => {
-      // K_ν is even in ν: K_{-n}(x) = K_n(x); negative-integer dispatch must abs() the order
-      for (const x of [1, 5]) {
-        assert.strictEqual(special.besselKnu(-1, x), special.besselK(1, x))
-        assert.strictEqual(special.besselKnu(-2, x), special.besselK(2, x))
-        assert.strictEqual(special.besselKnu(-3, x), special.besselK(3, x))
-      }
-    })
-
-    it('should match mpmath at the exact issue #1361 point (nu comparable to x, just past the x=6 crossover)', () => {
-      // mpmath mp.dps=50: besselk(4.82, 7.18) -> 0.001539804158137713
-      // Previously returned ~0.000358 (the bare _KAsymptotic leading term, ~77% relative error)
-      // because the unconditional x>6 dispatch to _KAsymptotic ignored how nu compared to x.
-      // x=7.18 is just past the crossover, where _KAsymptotic's own optimal-truncation error
-      // (not machine precision) is the limiting factor -- measured ~4.6e-9 relative error here,
-      // the same mechanism precision-refs-special.py's "just past x=6 crossover" tolerance
-      // bucket already documents for besselK/besselKnu (tol=2e-6 there); precision=8 gives
-      // >2x headroom over the measured value.
-      assert(equal(special.besselKnu(4.82, 7.18), 0.001539804158137713, 8))
-    })
-
-    it('should match mpmath for an order requiring multiple order-reduction recurrence steps, on both sides of the x=6 crossover', () => {
-      // mpmath mp.dps=50: besselk(7.5, 5.0) -> 0.39645666605558294 (x<=6 connection-formula base)
-      // mpmath mp.dps=50: besselk(7.5, 10.0) -> 0.00023814095655825686 (x>6 asymptotic base)
-      // nu=7.5 reduces to mu=-0.5, n=8: 7 upward-recurrence steps from the base pair. Default
-      // precision=10 (1e-10) holds comfortably here: measured relative error is ~1.06e-11 at
-      // x=5.0 (float rounding through the recurrence) and exactly 0 at x=10.0 -- nu=7.5 is
-      // half-integer, so both the connection formula and _KAsymptotic terminate exactly.
-      assert(equal(special.besselKnu(7.5, 5.0), 0.39645666605558294))
-      assert(equal(special.besselKnu(7.5, 10.0), 0.00023814095655825686))
-    })
-
-    it('should be even in nu (K_{-nu} = K_nu) for a fractional order in the fixed regime', () => {
-      // The order-reduction fix reduces on Math.abs(nu); this exercises that reduction still
-      // produces the correct value (not just internal self-consistency) for negative nu.
-      // mpmath mp.dps=50: besselk(-7.5, 10.0) -> 0.00023814095655825686 (same as +7.5)
-      assert(equal(special.besselKnu(-7.5, 10.0), 0.00023814095655825686))
-      assert.strictEqual(special.besselKnu(-7.5, 10.0), special.besselKnu(7.5, 10.0))
-    })
-
-    it('should throw for non-finite nu instead of looping forever', () => {
-      // Math.round(Math.abs(nu)) would be Infinity for nu = +-Infinity, making
-      // _besselKnuReduced's upward-recurrence loop (bound by that value) run forever instead
-      // of terminating after a bounded number of steps.
-      for (const nu of [Infinity, -Infinity, NaN]) {
-        assert.throws(() => special.besselKnu(nu, 5))
-      }
-    })
-  })
-
-  describe('.beta()', () => {
-    it('should return exact values for small positive integer arguments', () => {
-      // B(1,1) = 1 exactly
-      assert.strictEqual(special.beta(1, 1), 1)
-      // B(1,4) = 1/4 = 0.25 exactly — the YuleSimon quantile bug case
-      assert.strictEqual(special.beta(1, 4), 0.25)
-      // Symmetry: B(4,1) must equal B(1,4)
-      assert.strictEqual(special.beta(4, 1), 0.25)
-      // B(2,3) = 1/12; not exactly representable but must be closest double
-      assert(Math.abs(special.beta(2, 3) - 1 / 12) <= Number.EPSILON * (1 / 12))
-      // B(3,3) = 1/30; same criterion
-      assert(Math.abs(special.beta(3, 3) - 1 / 30) <= Number.EPSILON * (1 / 30))
-      // B(3,5) = 2!·4!/7! = 1/105; verify correct value and symmetry
-      assert(Math.abs(special.beta(3, 5) - 1 / 105) <= 2 * Number.EPSILON * (1 / 105))
-      assert(special.beta(5, 3) === special.beta(3, 5))
-    })
-
-    it('should agree with the logGamma path for non-integer arguments', () => {
-      // B(0.5, 0.5) = pi; non-integer path unchanged
-      assert(equal(special.beta(0.5, 0.5), Math.PI))
-      // B(1.5, 2.5) = Γ(1.5)Γ(2.5)/Γ(4) = (√π/2)(3√π/4)/6 = π/16
-      assert(equal(special.beta(1.5, 2.5), Math.PI / 16))
-    })
-
-    it('should fall back to logGamma when min(x,y) > 30', () => {
-      // beta(31, 31) has min=31 > 30, triggering the logGamma path
-      const v = special.beta(31, 31)
-      assert(Math.abs(v - 1.3861667124709484e-19) < 1e-30)
-    })
-  })
-
-  describe('.betaIncomplete()', () => {
-    it('B(a, b, x) should be equal to 0 if b > 0 and x <= 0', () => {
-      for (const a of [0.5, 1, 2]) {
-        for (const b of [1, 2]) {
-          assert(special.betaIncomplete(a, b, -1) === 0)
-        }
-      }
-    })
-
-    it('B(a, b, x) should be equal to B(a,b) if b > 0 and x >= 1', () => {
-      for (const a of [0.5, 1, 2]) {
-        for (const b of [1, 2]) {
-          const expected = Math.exp(special.logGamma(a) + special.logGamma(b) - special.logGamma(a + b))
-          assert(special.betaIncomplete(a, b, 2) === expected)
-        }
-      }
-    })
-
-    it('should return mpmath reference values near x=0 with mismatched a/b ratios', () => {
-      // Forward branch (x < (a+1)/(a+b+2)); precision probe at extreme small-x / high-b cases
-      // mpmath mp.dps=50: betainc(a, b, 0, x)
-      assert(equal(special.betaIncomplete(0.1, 10, 0.001), 5.007780291668116, 14))
-      assert(equal(special.betaIncomplete(0.01, 100, 0.0001), 91.19216638673353, 14))
-    })
-
-    it('should return mpmath reference values near x=1 with mismatched a/b ratios', () => {
-      // Backward branch (x > (a+1)/(a+b+2)); uses complement B(a,b) - B(b,a,1-x)
-      // mpmath mp.dps=50: betainc(a, b, 0, x)
-      // Full precision (14 digits): moderate a/b ratio, x not at extreme boundary
-      assert(equal(special.betaIncomplete(10, 0.5, 0.999), 0.5044901143493632, 14))
-      // Reduced precision (13 digits): a/b=100:1, mild boundary
-      assert(equal(special.betaIncomplete(10, 0.1, 0.95), 0.45383747791410134, 13))
-      // Reduced precision (11 digits): a/b=10000:1, x very close to 1 — complement B(a,b)-B(b,a,1-x)
-      // loses ~2 digits from catastrophic cancellation; inherent to the algorithm at this regime
-      assert(equal(special.betaIncomplete(100, 0.01, 0.9999), 3.7699233945564345, 11))
-    })
-
-    it('should return mpmath reference values at comfortable interior points', () => {
-      // mpmath mp.dps=50: betainc(a, b, 0, x)
-      assert(equal(special.betaIncomplete(2, 3, 0.4), 0.04373333333333334, 14))
-      assert(equal(special.betaIncomplete(5, 5, 0.5), 0.0007936507936507937, 14))
-    })
-
-    it('should return mpmath reference values for large a/b parameters', () => {
-      // mpmath mp.dps=50: betainc(a, b, 0, x); result is very small due to large shape params
-      assert(equal(special.betaIncomplete(50, 50, 0.5), 1.982330604283668e-31, 13))
-      // B(100,1,x) = x^100/100 analytically; at x=0.1 this is exactly 1e-102
-      assert(equal(special.betaIncomplete(100, 1, 0.1), 1e-102, 13))
-    })
-  })
-
-  describe('.f11()', () => {
-    const checkBesselIdentity = (aVals, zVals) => {
-      for (const a of aVals) {
-        for (const z of zVals) {
-          assert(equal(
-            special.f11(a, 2 * a, z),
-            Math.exp(z / 2 + (0.5 - a) * Math.log(z / 4) + special.logGamma(a + 0.5)) * special.besselInu(a - 0.5, z / 2)
-          ))
-        }
-      }
-    }
-
-    const checkF11Recurrence = (aVals, bVals, zVals) => {
-      for (const a of aVals) {
-        for (const b of bVals) {
-          for (const z of zVals) {
-            assert(equal(
-              a * special.f11(a + 1, b, z),
-              (b - a) * special.f11(a - 1, b, z) + (2 * a - b + z) * special.f11(a, b, z)
-            ))
-          }
-        }
-      }
-    }
-
-    describe('|z| < 50', () => {
-      it('f11(0, b, z) = 1', () => {
-        for (const b of [0.5, 1, 2]) {
-          for (const z of [0, 1, 10, 40]) {
-            assert(equal(special.f11(0, b, z), 1))
-          }
-        }
-      })
-
-      it('f11(b, b, z) = exp(z)', () => {
-        for (const b of [0.5, 1, 2]) {
-          for (const z of [0, 1, 10, 40]) {
-            assert(equal(special.f11(b, b, z), Math.exp(z)))
-          }
-        }
-      })
-
-      it('f11(2, 1, z) = (1 + z) * exp(z)', () => {
-        for (const z of [0, 1, 10, 40]) {
-          assert(equal(special.f11(2, 1, z), (1 + z) * Math.exp(z)))
-        }
-      })
-
-      it('f11(1, 2, z) = (exp(z) - 1) / z', () => {
-        for (const z of [1, 10, 40]) {
-          assert(equal(special.f11(1, 2, z), (Math.exp(z) - 1) / z))
-        }
-      })
-
-      it('(2z / sqrt(pi)) * f11(0.5, 1.5, -z^2) = erf(z)', () => {
-        for (const z of [0.1, 0.5, 1]) {
-          assert(equal(2 * z * special.f11(0.5, 1.5, -z * z) / Math.sqrt(Math.PI), special.erf(z)))
-        }
-      })
-
-      it('f11(a, 2a, z) = exp(z/2) (z/4)^(0.5 - a) gamma(a + 0.5) I(a - 0.5; z/2)', () => {
-        checkBesselIdentity([0.5, 1, 2], [1, 10, 40])
-      })
-
-      it('a f11(a+1, b, z) = (b - a) f11(a-1, b, z) + (2a - b + z) f11(a, b, z)', () => {
-        checkF11Recurrence([0.5, 1, 2], [0.5, 1, 2], [0, 1, 10, 40])
-      })
-
-      it('f11(1, 2, 49) = (exp(49) - 1) / 49 at 12-digit precision', () => {
-        // Taylor series requires ~115 terms to converge at z=49; would fail at 10-digit precision with MAX_ITER=100.
-        assert(equal(special.f11(1, 2, 49), (Math.exp(49) - 1) / 49, 12))
-      })
-    })
-
-    describe('|z| >= 50', () => {
-      it('f11(0, b, z) = 1', () => {
-        for (const b of [0.5, 1, 2]) {
-          for (const z of [50, 60, 75, 90]) {
-            assert(equal(special.f11(0, b, z), 1))
-          }
-        }
-      })
-
-      it('f11(b, b, z) = exp(z)', () => {
-        for (const b of [0.5, 1, 2]) {
-          for (const z of [50, 60, 75, 90]) {
-            assert(equal(special.f11(b, b, z), Math.exp(z)))
-          }
-        }
-      })
-
-      it('f11(2, 1, z) = (1 + z) * exp(z)', () => {
-        for (const z of [50, 60, 75, 90]) {
-          assert(equal(special.f11(2, 1, z), (1 + z) * Math.exp(z)))
-        }
-      })
-
-      it('f11(1, 2, z) = (exp(z) - 1) / z', () => {
-        for (const z of [50, 60, 75, 90]) {
-          assert(equal(special.f11(1, 2, z), (Math.exp(z) - 1) / z))
-        }
-      })
-
-      it('f11(a, 2a, z) = exp(z/2) (z/4)^(0.5 - a) gamma(a + 0.5) I(a - 0.5; z/2)', () => {
-        checkBesselIdentity([0.5, 1, 2], [50, 60, 75, 90])
-      })
-
-      it('a * f11(a+1, b, z) = (b - a) * f11(a-1, b, z) + (2a - b + z) * f11(a, b, z)', () => {
-        checkF11Recurrence([3, 7, 13], [3, 7, 13], [50, 60, 75, 90])
-      })
-
-      it('f11(1, 2, 50) = (exp(50) - 1) / 50', () => {
-        assert(equal(special.f11(1, 2, 50), (Math.exp(50) - 1) / 50))
-      })
-
-      it('f11(1, 2, 100) = (exp(100) - 1) / 100', () => {
-        assert(equal(special.f11(1, 2, 100), (Math.exp(100) - 1) / 100))
-      })
-    })
-  })
-
-  describe('.gamma(), .logGamma()', () => {
-    it('logGamma(z) = ln(gamma(z))', () => {
-      for (const x of [0.1, 0.5, 1, 1.5, 2, 3, 5, 10, 20, 50, 99.5]) {
-        const g = special.gamma(x)
-        const lng = special.logGamma(x)
-        const err = Math.abs(Math.log(g) - lng)
-        // when lng === 0 (x = 1, 2) relative error is undefined; check absolute instead
-        assert(Math.abs(lng) < 1e-10 ? err < 1e-14 : err / Math.abs(lng) < 0.01)
-      }
-    })
-
-    describe('.gamma()', () => {
-      it('should return reference values (mpmath mp.dps=50)', () => {
-        // Positive baseline plus the reflection branch at negative half-integers.
-        assert(equal(special.gamma(0.5), 1.772453850905516, 14))
-        assert(equal(special.gamma(-0.5), -3.544907701811032, 14))
-        assert(equal(special.gamma(-1.5), 2.363271801207355, 14))
-        assert(equal(special.gamma(-2.5), -0.9453087204829419, 14))
-      })
-
-      it('should return Infinity at the non-positive integer poles', () => {
-        // ADR-0015: divergence returns Infinity specifically (not NaN, not a huge finite).
-        assert.strictEqual(special.gamma(0), Infinity)
-        assert.strictEqual(special.gamma(-1), Infinity)
-        assert.strictEqual(special.gamma(-2), Infinity)
-      })
-
-      it('should stay full-precision within 1e-6 of a negative integer pole', () => {
-        // mpmath mp.dps=60 evaluated at the exact double of (-1+1e-7) / (-2+1e-7); the
-        // (-1)^n-signed reduced sin(πz) must not lose the fractional offset.
-        assert(equal(special.gamma(-1 + 1e-7), -10000000.428048076, 13))
-        assert(equal(special.gamma(-2 + 1e-7), 5000000.458472761, 13))
-      })
-    })
-
-    describe('.logGamma()', () => {
-      it('should return reference values ln|Γ(z)| (mpmath mp.dps=50)', () => {
-        // Positive baseline (direct Lanczos path) and the log-reflection branch at negative
-        // half-integers. The g=7.5 9-term Lanczos achieves ~1e-15, so tolerance is 1e-14.
-        // z = 0.25 exercises the 0 < z < 0.5 positive-reflection branch.
-        assert(equal(special.logGamma(0.25), 1.2880225246980774, 14))
-        assert(equal(special.logGamma(0.5), 0.5723649429247001, 14))
-        assert(equal(special.logGamma(1.5), -0.12078223763524522, 14))
-        assert(equal(special.logGamma(2.5), 0.2846828704729192, 14))
-        // ln(2!) and ln(4!): pinned as mpmath literals, not Math.log() calls
-        assert(equal(special.logGamma(3), 0.6931471805599453, 14))
-        assert(equal(special.logGamma(5), 3.1780538303479458, 14))
-        assert(equal(special.logGamma(10), 12.801827480081469, 14))
-        assert(equal(special.logGamma(-0.5), 1.2655121234846454, 14))
-        assert(equal(special.logGamma(-1.5), 0.860047015376481, 14))
-        // z = -2.5: reflection subtracts log(π) - logGamma(3.5) ≈ 1.14 - 1.20; the ~5%
-        // cancellation costs ~1 digit, limiting accuracy to ~2e-14 for this argument.
-        assert(equal(special.logGamma(-2.5), -0.056243716497674054, 13))
-      })
-
-      it('should return Infinity at the non-positive integer poles', () => {
-        // ADR-0015: divergence returns Infinity specifically (not NaN, not a huge finite).
-        assert.strictEqual(special.logGamma(0), Infinity)
-        assert.strictEqual(special.logGamma(-1), Infinity)
-        assert.strictEqual(special.logGamma(-2), Infinity)
-      })
-
-      it('should stay full-precision within 1e-6 of a negative integer pole', () => {
-        // mpmath mp.dps=60 at the exact double of (-1+1e-7) / (-2+1e-7); |sin(πz)| reduced mod π keeps the offset.
-        assert(equal(special.logGamma(-1 + 1e-7), 16.118095693763127, 14))
-        assert(equal(special.logGamma(-2 + 1e-7), 15.424948562092922, 14))
-      })
-    })
-  })
-
-  describe('.gammaLowerIncomplete(), .gammaUpperIncomplete()', () => {
-    it('should vanish below 0', () => {
-      for (const s of [0.5, 1, 2, 5, 10]) {
-        assert(special.gammaLowerIncomplete(s, -10) === 0)
-      }
-    })
-
-    // Reference values from mpmath at mp.dps=70.
-    it('should match mpmath reference values in the deep series branch (x << a)', () => {
-      // Series expansion path: x < s+1
-      assert(equal(special.gammaLowerIncomplete(5, 0.5), 1.7211562995584078e-4))
-      assert(equal(special.gammaLowerIncomplete(10, 1.0), 1.1142547833872067e-7))
-      assert(equal(special.gammaLowerIncomplete(50, 5.0), 2.1810592140784887e-32))
-      // Complementary upper values
-      assert(equal(special.gammaUpperIncomplete(5, 0.5), 0.9998278843700441))
-      assert(equal(special.gammaUpperIncomplete(10, 1.0), 0.9999998885745217))
-    })
-
-    it('should match mpmath reference values in the crossover region (x near a)', () => {
-      // Near x = a the series and CF both converge most slowly; the implementation uses
-      // the series branch (x < s+1) here.
-      assert(equal(special.gammaLowerIncomplete(5, 5.0), 0.5595067149347875))
-      assert(equal(special.gammaUpperIncomplete(5, 5.0), 0.4404932850652124))
-      assert(equal(special.gammaLowerIncomplete(10, 10.0), 0.5420702855281478))
-      assert(equal(special.gammaUpperIncomplete(10, 10.0), 0.4579297144718522))
-      assert(equal(special.gammaLowerIncomplete(50, 50.0), 0.5188083154720433))
-      assert(equal(special.gammaUpperIncomplete(50, 50.0), 0.48119168452795674))
-    })
-
-    it('should match mpmath reference values in the deep CF branch (x >> a)', () => {
-      // Continued-fraction path: x >= s+1; upper incomplete gamma computed directly.
-      assert(equal(special.gammaUpperIncomplete(1, 20.0), 2.061153622438558e-9))
-      assert(equal(special.gammaUpperIncomplete(5, 50.0), 5.4497019829205295e-17, 10))
-      assert(equal(special.gammaUpperIncomplete(10, 100.0), 1.1253473960842733e-31, 10))
-    })
-
-    it('should match mpmath reference values for large a', () => {
-      // Large shape parameter; series branch is used (x = a < a+1).
-      assert(equal(special.gammaLowerIncomplete(100, 100.0), 0.5132987982791487))
-      assert(equal(special.gammaUpperIncomplete(100, 100.0), 0.48670120172085135))
-      assert(equal(special.gammaLowerIncomplete(1000, 1000.0), 0.5042052441802155))
-      assert(equal(special.gammaUpperIncomplete(1000, 1000.0), 0.4957947558197845))
-    })
-
-    // mpmath mp.dps=50: mp.gammainc(s, x, mp.inf, regularized=True). CF branch (x >= s+1)
-    // with s and x both large and close together -- the near-diagonal regime where _gui's
-    // MAX_ITER=100 cap silently truncated before convergence and its shared prefactor with
-    // _gli cancelled two ~4e4-magnitude terms down to an O(1) result (#1348).
-    it('should match mpmath reference values in the near-diagonal large-s CF branch', () => {
-      assert(equal(special.gammaUpperIncomplete(4989, 5000), 0.436307027364574, 13))
-      assert(equal(special.gammaUpperIncomplete(4995, 5000), 0.46993291332034903, 13))
-      assert(equal(special.gammaUpperIncomplete(4998, 5000), 0.48683689071721553, 13))
-    })
-
-    it('P + Q should equal 1 for all algorithm regions', () => {
-      const pairs = [
-        [5, 0.5], [10, 1.0], [50, 5.0],
-        [5, 5.0], [10, 10.0], [50, 50.0],
-        [1, 20.0], [5, 50.0], [10, 100.0],
-        [100, 100.0], [1000, 1000.0]
-      ]
-      for (const [s, x] of pairs) {
-        const p = special.gammaLowerIncomplete(s, x)
-        const q = special.gammaUpperIncomplete(s, x)
-        assert(equal(p + q, 1))
-      }
-    })
-
-    it('should converge to gamma(s) as x -> inf', () => {
-      for (const s of [0.5, 1, 2, 5, 10]) {
-        assert(equal(special.gammaLowerIncomplete(s, 1e5), 1))
-      }
-    })
-  })
-
-  describe('.gammaLowerIncompleteInv()', () => {
-    it('should return 0 for p = 0', () => {
-      for (const a of [1, 2, 5, 10, 20]) {
-        assert(special.gammaLowerIncompleteInv(a, 0) === 0)
-      }
-    })
-
-    it('should return Infinity for p = 1', () => {
-      for (const a of [1, 2, 5, 10, 20]) {
-        assert(special.gammaLowerIncompleteInv(a, 1) === Infinity)
-      }
-    })
-
-    it('should round-trip with gammaLowerIncomplete for a >= 1', () => {
-      for (const a of [1, 2, 5, 10, 20]) {
-        for (const p of [0.01, 0.1, 0.5, 0.9, 0.99]) {
-          const x = special.gammaLowerIncompleteInv(a, p)
-          assert(Math.abs(special.gammaLowerIncomplete(a, x) - p) < 1e-10)
-        }
-      }
-    })
-
-    it('should round-trip with gammaLowerIncomplete for a < 1', () => {
-      for (const a of [0.1, 0.3, 0.5, 0.7, 0.9]) {
-        for (const p of [0.01, 0.1, 0.5, 0.9, 0.99]) {
-          const x = special.gammaLowerIncompleteInv(a, p)
-          assert(Math.abs(special.gammaLowerIncomplete(a, x) - p) < 1e-10)
-        }
-      }
-    })
-
-    it('should return a known value: gammaLowerIncompleteInv(2, 0.5) ≈ 1.6783', () => {
-      // gammaLowerIncomplete(2, 1.6783469900166612) = 0.5 exactly
-      assert(Math.abs(special.gammaLowerIncompleteInv(2, 0.5) - 1.6783469900166612) < 1e-10)
-    })
-
-    it('should handle extreme lower tail (small p)', () => {
-      const x = special.gammaLowerIncompleteInv(5, 1e-10)
-      assert(isFinite(x) && x > 0)
-      assert(Math.abs(special.gammaLowerIncomplete(5, x) - 1e-10) < 1e-15)
-    })
-
-    it('should converge for very small p (p = 1e-30, a = 1)', () => {
-      // mpmath: -log(1 - 1e-30) ≈ 1e-30; round-trip must recover p to relative 1e-10
-      const x = special.gammaLowerIncompleteInv(1, 1e-30)
-      assert(isFinite(x) && x > 0)
-      assert(Math.abs(special.gammaLowerIncomplete(1, x) - 1e-30) / 1e-30 < 1e-10)
-    })
-
-    it('should converge for p = 1e-31, a = 0.1 where x_true < 1e-300', () => {
-      // Leading-term inversion gives x_true ≈ (p * Gamma(1.1))^10 ≈ 6e-311, below 1e-300.
-      // The old absolute 1e-300 floor clamped the iterate upward and stalled convergence.
-      const x = special.gammaLowerIncompleteInv(0.1, 1e-31)
-      assert(isFinite(x) && x > 0 && x < 1e-300)
-      assert(Math.abs(special.gammaLowerIncomplete(0.1, x) - 1e-31) / 1e-31 < 1e-10)
-    })
-
-    it('should handle extreme upper tail (p close to 1)', () => {
-      const x = special.gammaLowerIncompleteInv(5, 1 - 1e-10)
-      assert(isFinite(x) && x > 0)
-      assert(Math.abs(special.gammaLowerIncomplete(5, x) - (1 - 1e-10)) < 1e-10)
-    })
-  })
-
-  describe('.hurwitzZeta(), .riemannZeta()', () => {
-    it('hurwitzZeta should return Infinity at the pole s = 1', () => {
-      assert(special.hurwitzZeta(1, 1) === Infinity)
-      assert(special.hurwitzZeta(1, 2) === Infinity)
-      assert(special.hurwitzZeta(1, 0.5) === Infinity)
-      // Values just outside the EPS guard must be large but finite
-      assert(isFinite(special.hurwitzZeta(1 + 1e-10, 1)))
-      assert(special.hurwitzZeta(1 + 1e-10, 1) > 1e9)
-    })
-
-    it('hurwitzZeta should be accurate for s near 1', () => {
-      // Reference values from the 3-term Laurent ζ(s) = 1/(s-1) + γ₀ − γ₁(s−1) + γ₂(s−1)²/2
-      // (DLMF 25.2.8) using γ₀=0.5772156649, γ₁=−0.0728158455, γ₂=−0.0096903632.
-      // At s=1.01 the 4th-term error is ~3e-12; tolerances are set well above that.
-      assert(equal(special.hurwitzZeta(1.01, 1), 100.577943338838, 10))
-      assert(equal(special.hurwitzZeta(1.05, 1), 20.580844344222, 8))
-      assert(equal(special.hurwitzZeta(1.1, 1), 10.584448797634, 7))
-      // ζ(s, 2) = ζ(s, 1) − 1^{-s} = ζ(s) − 1; exercises a ≠ 1 with n-formula unchanged
-      assert(equal(special.hurwitzZeta(1.05, 2), 20.580844344222 - 1, 8))
-    })
-
-    it('hurwitzZeta should satisfy the recurrence ζ(s, a) = a^{-s} + ζ(s, a+1) for s near 1', () => {
-      // Verifies that the corrected partial-sum length produces consistent values across the
-      // recurrence, which the fixed-point tests at s=1.01/1.05/1.1 alone cannot catch.
-      for (const s of [1.05, 1.1, 1.25, 1.5]) {
-        for (const a of [0.5, 1, 2, 4]) {
-          assert(equal(
-            special.hurwitzZeta(s, a),
-            Math.pow(a, -s) + special.hurwitzZeta(s, a + 1),
-            6
-          ))
-        }
-      }
-    })
-
-    it('riemannZeta(s) - hurwitzZeta(s, n+1) = H(s, n)', () => {
-      for (const s of [1.01, 1.1, 2, 5, 10]) {
-        let sum = 0
-        for (let n = 1; n < 100; n++) {
-          sum += 1 / Math.pow(n, s)
-          assert(Math.abs(sum - special.riemannZeta(s) + special.hurwitzZeta(s, n + 1)) / sum < 1e-6)
-        }
-      }
-    })
-
-    it('should return Infinity at the pole s = 1', () => {
-      assert(special.riemannZeta(1) === Infinity)
-    })
-
-    it('should be accurate near s = 1 via Laurent expansion', () => {
-      // Reference values from five-term Laurent expansion (DLMF 25.2.4, γ₀–γ₄); truncation error O(d^5)
-      // s > 1 side
-      assert(Math.abs(special.riemannZeta(1.0001) / 10000.577222946486 - 1) < 1e-8)
-      assert(Math.abs(special.riemannZeta(1.001) / 1000.5772884762018 - 1) < 1e-8)
-      assert(Math.abs(special.riemannZeta(1.01) / 100.5779433388382 - 1) < 1e-8)
-      assert(Math.abs(special.riemannZeta(1.02) / 50.5786700377986 - 1) < 1e-8)
-      // s=1.05 and s=1.1 cross-checked against independent hurwitzZeta(s, 1) references (ζ(s) = ζ(s,1))
-      assert(Math.abs(special.riemannZeta(1.05) / 20.580844344222 - 1) < 1e-8)
-      assert(Math.abs(special.riemannZeta(1.1) / 10.584448465 - 1) < 1e-8)
-      // s < 1 side (Laurent branch fires for |s-1| < 0.1 in both directions)
-      assert(Math.abs(special.riemannZeta(0.999) / (-999.422857150944) - 1) < 1e-8)
-    })
-  })
-
-  describe('.lambertW0()', () => {
-    it('should return NaN for z < -1/e', () => {
-      assert(isNaN(special.lambertW0(-1)))
-      assert(isNaN(special.lambertW0(-0.5)))
-      assert(isNaN(special.lambertW0(-Math.exp(-1) - 1e-10)))
-    })
-
-    it('should return -1 at the branch point z = -1/e', () => {
-      assert(Math.abs(special.lambertW0(-Math.exp(-1)) + 1) < 1e-6)
-    })
-
-    it('should return 0 at z = 0', () => {
-      assert(special.lambertW0(0) === 0)
-    })
-
-    it('should return 1 at z = e', () => {
-      assert(equal(special.lambertW0(Math.E), 1))
-    })
-
-    it('should satisfy the W * exp(W) = x equation for x >= 0', () => {
-      for (const x of [0, 1e-10, 0.1, 1, Math.E, 5, 10]) {
-        const w = special.lambertW0(x)
-        assert(equal(w * Math.exp(w), x))
-      }
-    })
-
-    it('should satisfy the W * exp(W) = x equation for x in [-1/e, 0)', () => {
-      for (const x of [-1e-6, -0.1, -0.2, -1 / Math.E + 1e-10]) {
-        const w = special.lambertW0(x)
-        assert(equal(w * Math.exp(w), x))
-      }
-    })
-  })
-
-  describe('.lambertW1m()', () => {
-    it('should return NaN for z < -1/e', () => {
-      assert(isNaN(special.lambertW1m(-1)))
-      assert(isNaN(special.lambertW1m(-0.5)))
-      assert(isNaN(special.lambertW1m(-Math.exp(-1) - 1e-10)))
-    })
-
-    it('should return NaN for z >= 0', () => {
-      assert(isNaN(special.lambertW1m(0)))
-      assert(isNaN(special.lambertW1m(1)))
-      assert(isNaN(special.lambertW1m(0.1)))
-    })
-
-    it('should return -1 at the branch point z = -1/e', () => {
-      assert(Math.abs(special.lambertW1m(-Math.exp(-1)) + 1) < 1e-6)
-    })
-
-    it('should return known value at z = -0.1', () => {
-      assert(equal(special.lambertW1m(-0.1), -3.577152063957297))
-    })
-
-    it('should satisfy the W * exp(W) = x equation', () => {
-      for (const x of [-1e-6, -0.1, -0.2, -1 / Math.E + 1e-10]) {
-        const w = special.lambertW1m(x)
-        assert(equal(w * Math.exp(w), x))
-      }
-    })
-
-    it('should return known values near the branch cut', () => {
-      assert(equal(special.lambertW1m(-0.2), -2.5426413577735265))
-      assert(equal(special.lambertW1m(-0.05), -4.499755288523487))
-    })
-
-    it('should satisfy W * exp(W) = z near the branch cut (z in [-1/e, -0.1])', () => {
-      for (const x of [-0.15, -0.2, -0.3, -1 / Math.E + 1e-10]) {
-        const w = special.lambertW1m(x)
-        assert(equal(w * Math.exp(w), x))
-      }
-    })
-  })
-
-  describe('.marcumQ()', () => {
-    const check = (x, y, mu) => {
-      const q1 = special.marcumQ(mu + 1, x, y)
-      const q2 = special.marcumQ(mu, x, y)
-      const q3 = special.marcumQ(mu + 2, x, y)
-      const q4 = special.marcumQ(mu - 1, x, y)
-      const r = x > mu
-        ? ((x - mu) * q1 + (y + mu) * q2) / (x * q3 + y * q4)
-        : ((y + mu) * q2) / (x * q3 + (mu - x) * q1 + y * q4)
-      assert(equal(r, 1))
-    }
-
-    describe('special cases', () => {
-      describe('x = 0', () => {
-        it('should satisfy the recurrence relation', () => {
-          for (const x of [1, 10, 30]) {
-            for (const mu of [2, 4, 7]) {
-              assert(equal(special.marcumQ(mu, x, 0), 1))
-            }
-          }
-        })
-      })
-
-      describe('y = 1', () => {
-        it('should satisfy the recurrence relation', () => {
-          for (const y of [40, 60, 80, 100]) {
-            for (const mu of [2, 4, 7]) {
-              assert(equal(special.marcumQ(mu, 0, y), special.gammaUpperIncomplete(mu, y)))
-            }
-          }
-        })
-      })
-    })
-
-    describe('series expansion', () => {
-      describe('Q', () => {
-        it('should satisfy the recurrence relation', () => {
-          for (const [x, y, mu] of [[1, 40, 2], [10, 60, 4], [29, 99, 7]]) {
-            check(x, y, mu)
-          }
-        })
-      })
-
-      describe('P', () => {
-        it('should satisfy the recurrence relation', () => {
-          for (const [x, y, mu] of [[1, 10, 30], [10, 15, 32], [29, 20, 35]]) {
-            check(x, y, mu)
-          }
-        })
-      })
-
-      it('should match scipy ncx2 reference values', () => {
-        [
-          { mu: 4, x: 10, y: 20, p: 0.8840146502552253, q: 0.11598534974477473 },
-          { mu: 2, x: 5, y: 3, p: 0.10329898661024607, q: 0.8967010133897539 }
-        ].forEach(d => {
-          assert(equal(special.marcumQ(d.mu, d.x, d.y), d.q), `marcumQ(${d.mu}, ${d.x}, ${d.y})`)
-          assert(equal(special.marcumP(d.mu, d.x, d.y), d.p), `marcumP(${d.mu}, ${d.x}, ${d.y})`)
-        })
-      })
-    })
-
-    describe('asymptotic expansion for large xi', () => {
-      describe('Q', () => {
-        it('should satisfy the recurrence relation', () => {
-          for (const [x, y, mu] of [[35, 60, 3], [70, 100, 4.5], [134, 160, 6]]) {
-            const q1 = special.marcumQ(mu + 1, x, y)
-            const q2 = special.marcumQ(mu, x, y)
-            const q3 = special.marcumQ(mu + 2, x, y)
-            const q4 = special.marcumQ(mu - 1, x, y)
-
-            assert(equal(((x - mu) * q1 + (y + mu) * q2) / (x * q3 + y * q4), 1))
-          }
-        })
-      })
-
-      describe('P', () => {
-        it('should satisfy the recurrence relation', () => {
-          for (const [x, y, mu] of [[45, 30, 3], [90, 70, 4.5], [134, 110, 6]]) {
-            const q1 = special.marcumQ(mu + 1, x, y)
-            const q2 = special.marcumQ(mu, x, y)
-            const q3 = special.marcumQ(mu + 2, x, y)
-            const q4 = special.marcumQ(mu - 1, x, y)
-
-            assert(equal(((x - mu) * q1 + (y + mu) * q2) / (x * q3 + y * q4), 1))
-          }
-        })
-      })
-
-      it('should match scipy ncx2 reference values', () => {
-        [
-          { mu: 5, x: 40, y: 60, q: 0.05987990370344369, p: 0.9401200962965562 },
-          { mu: 3, x: 120, y: 150, q: 0.04687422045974286, p: 0.9531257795402576 },
-          { mu: 10, x: 70, y: 45, q: 0.9994063950877903, p: 0.0005936049122101624 },
-          { mu: 3, x: 46, y: 5, q: 0.9999999999966671, p: 3.3330073779888013e-12 }
-        ].forEach(d => {
-          assert(equal(special.marcumQ(d.mu, d.x, d.y), d.q), `marcumQ(${d.mu}, ${d.x}, ${d.y})`)
-          assert(equal(special.marcumP(d.mu, d.x, d.y), d.p), `marcumP(${d.mu}, ${d.x}, ${d.y})`)
-        })
-      })
-    })
-
-    describe('quadrature', () => {
-      it('should satisfy the recurrence relation', () => {
-        for (const [x, y, mu] of [[40, 0.5, 3], [60, 1, 5], [79, 2, 8]]) {
-          const q1 = special.marcumQ(mu + 1, x, y)
-          const q2 = special.marcumQ(mu, x, y)
-          const q3 = special.marcumQ(mu + 2, x, y)
-          const q4 = special.marcumQ(mu - 1, x, y)
-
-          assert(equal(((x - mu) * q1 + (y + mu) * q2) / (x * q3 + y * q4), 1))
-        }
-      })
-
-      it('should match scipy ncx2 reference values', () => {
-        [
-          { mu: 140, x: 100, y: 235, p: 0.4015210444334114, q: 0.5984789555665886 },
-          { mu: 140, x: 100, y: 200, p: 0.011694400751604403, q: 0.9883055992483956 },
-          { mu: 60, x: 200, y: 400, p: 0.9999999966690462, q: 3.3309538428122277e-9 },
-          { mu: 5, x: 50, y: 4, p: 7.677483509552288e-16, q: 0.9999999999999992 },
-          { mu: 8, x: 60, y: 3, p: 3.3894189213527325e-23, q: 1.0 }
-        ].forEach(d => {
-          assert(equal(special.marcumQ(d.mu, d.x, d.y), d.q), `marcumQ(${d.mu}, ${d.x}, ${d.y})`)
-          assert(equal(special.marcumP(d.mu, d.x, d.y), d.p), `marcumP(${d.mu}, ${d.x}, ${d.y})`)
-        })
-      })
-
-      // #1179: ys = y/mu far below 1 catastrophically cancelled in _zetaxy(), producing
-      // NaN (or, if naively rationalized, a silently-wrong finite plateau) instead of the
-      // correct deep-lower-tail value. Reference values from mpmath at mp.dps=50 via the
-      // series P_mu(x,y) = e^-x * sum_n (x^n/n!) * P_(mu+n)(y) (Eq. 7 of the paper), which
-      // is cancellation-free for any y and independent of the ranjs implementation.
-      it('should match reference values for y << mu (deep lower tail, #1179)', () => {
-        [
-          // Below _pqTrap's underflow threshold (halfMuZeta2 > -log(DELTA)): the
-          // fixed zeta must stay finite so the shortcut branch fires cleanly instead
-          // of falling through to a NaN-tainted quadrature. Exact reproduction from #1179.
-          { mu: 1, x: 32, y: 2.4623e-32, p: 0, q: 1 },
-          { mu: 2.5, x: 40, y: 1e-20, p: 0, q: 1 },
-          // Above the underflow threshold: the quadrature itself executes with the
-          // fixed zeta, so these lock in correctness of _pqTrap's full path, not just
-          // the shortcut. mpmath dps=50: series Eq. 7.
-          { mu: 1, x: 32, y: 1e-16, p: 1.2664165549094195e-30, q: 1 },
-          { mu: 1, x: 32, y: 1e-10, p: 1.2664165568723631e-24, q: 1 },
-          { mu: 1, x: 32, y: 1e-4, p: 1.2683804484834176e-18, q: 1 },
-          // u = 4 * (x/mu) * (y/mu) = 0.4992, just below the u < 0.5 branch
-          // boundary in _zetaxy(): locks in that the rationalized d2 formula
-          // agrees with the unchanged u >= 0.5 branch near the switchover.
-          { mu: 1, x: 32, y: 0.0039, p: 5.243259585512871e-17, q: 1 }
-        ].forEach(d => {
-          assert(equal(special.marcumQ(d.mu, d.x, d.y), d.q), `marcumQ(${d.mu}, ${d.x}, ${d.y})`)
-          assert(equal(special.marcumP(d.mu, d.x, d.y), d.p), `marcumP(${d.mu}, ${d.x}, ${d.y})`)
-        })
-      })
-    })
-
-    describe('recurrence relation', () => {
-      it('should satisfy the recurrence relation', () => {
-        // Both x < mu and x > mu inside the transition band, exercising both
-        // forms of the recurrence test.
-        for (const [x, y, mu] of [[40, 120, 80], [55, 155, 100], [70, 190, 120]]) {
-          check(x, y, mu)
-        }
-        for (const [x, y, mu] of [[80, 120, 40], [110, 165, 55], [140, 210, 70]]) {
-          check(x, y, mu)
-        }
-      })
-
-      it('should match scipy ncx2 reference values', () => {
-        [
-          { mu: 100, x: 50, y: 150, p: 0.5117578749745552, q: 0.48824212502544484 },
-          { mu: 90, x: 60, y: 140, p: 0.2499150811158282, q: 0.7500849188841718 }
-        ].forEach(d => {
-          assert(equal(special.marcumQ(d.mu, d.x, d.y), d.q), `marcumQ(${d.mu}, ${d.x}, ${d.y})`)
-          assert(equal(special.marcumP(d.mu, d.x, d.y), d.p), `marcumP(${d.mu}, ${d.x}, ${d.y})`)
-        })
-      })
-    })
-
-    describe('large mu asymptotic', () => {
-      it('should satisfy the recurrence relation', () => {
-        // mu >= 135 is the dispatch threshold; at the boundary the mu-1 order
-        // may fall on the recurrence branch, but the three-term identity holds
-        // across any mix of correct branches. Both x < mu and x > mu run.
-        for (const [x, y, mu] of [[30, 165, 135], [55, 235, 180], [80, 330, 250]]) {
-          check(x, y, mu)
-        }
-        for (const [x, y, mu] of [[155, 290, 135], [200, 365, 165], [260, 460, 200]]) {
-          check(x, y, mu)
-        }
-      })
-
-      it('should match scipy ncx2 reference values', () => {
-        [
-          { mu: 135, x: 40, y: 170, p: 0.3755801225650498, q: 0.6244198774349506 },
-          { mu: 135, x: 44, y: 188, p: 0.7323059732316725, q: 0.26769402676832693 },
-          { mu: 150, x: 60, y: 205, p: 0.38907639626893925, q: 0.6109236037310609 },
-          { mu: 200, x: 55, y: 248, p: 0.35242643120412376, q: 0.6475735687958762 },
-          { mu: 160, x: 70, y: 250, p: 0.8745123498463065, q: 0.1254876501536935 },
-          { mu: 180, x: 50, y: 230, p: 0.5093676830927159, q: 0.49063231690728504 }
-        ].forEach(d => {
-          assert(equal(special.marcumQ(d.mu, d.x, d.y), d.q), `marcumQ(${d.mu}, ${d.x}, ${d.y})`)
-          assert(equal(special.marcumP(d.mu, d.x, d.y), d.p), `marcumP(${d.mu}, ${d.x}, ${d.y})`)
-        })
-      })
-    })
-  })
-
-  describe('.marcumP()', () => {
-    describe('special cases', () => {
-      it('should return 0 for y = 0', () => {
-        for (const [mu, x] of [[2, 0], [4, 15], [7, 30]]) {
-          assert(special.marcumP(mu, x, 0) === 0)
-        }
-      })
-
-      it('should equal the lower incomplete gamma for x = 0', () => {
-        for (const y of [40, 70, 100]) {
-          for (const mu of [2, 4, 7]) {
-            assert(equal(special.marcumP(mu, 0, y), special.gammaLowerIncomplete(mu, y)))
-          }
-        }
-      })
-    })
-
-    it('should satisfy the recurrence relation', () => {
-      const check = (x, y, mu) => {
-        const p1 = special.marcumP(mu + 1, x, y)
-        const p2 = special.marcumP(mu, x, y)
-        const p3 = special.marcumP(mu + 2, x, y)
-        const p4 = special.marcumP(mu - 1, x, y)
-        const r = x > mu
-          ? ((x - mu) * p1 + (y + mu) * p2) / (x * p3 + y * p4)
-          : ((y + mu) * p2) / (x * p3 + (mu - x) * p1 + y * p4)
-        assert(equal(r, 1))
-      }
-      // Series, asymptotic, quadrature, recurrence and large-mu regimes. The
-      // quadrature and recurrence points stay close enough to the transition
-      // that P is above the underflow limit, so the relation is meaningfully
-      // exercised.
-      for (const [x, y, mu] of [[1, 10, 30], [15, 15, 32], [29, 20, 35]]) {
-        check(x, y, mu)
-      }
-      for (const [x, y, mu] of [[45, 30, 3], [90, 70, 4.5], [134, 110, 6]]) {
-        check(x, y, mu)
-      }
-      for (const [x, y, mu] of [[40, 95, 80], [55, 126, 100], [70, 158, 120]]) {
-        check(x, y, mu)
-      }
-      for (const [x, y, mu] of [[40, 120, 80], [55, 155, 100], [70, 190, 120]]) {
-        check(x, y, mu)
-      }
-      for (const [x, y, mu] of [[30, 165, 135], [55, 235, 180], [80, 330, 245]]) {
-        check(x, y, mu)
-      }
-    })
-  })
-
-  describe('marcumQ and marcumP', () => {
-    it('should satisfy marcumQ + marcumP = 1 across all branches', () => {
-      const identity = (x, y, mu) => {
-        assert(equal(special.marcumQ(mu, x, y) + special.marcumP(mu, x, y), 1))
-      }
-      // Series, asymptotic, quadrature, recurrence and large-mu regimes.
-      for (const [x, y, mu] of [[1, 40, 2], [15, 70, 4], [29, 99, 7]]) {
-        identity(x, y, mu)
-      }
-      for (const [x, y, mu] of [[35, 60, 3], [70, 100, 4.5], [134, 160, 6]]) {
-        identity(x, y, mu)
-      }
-      for (const [x, y, mu] of [[40, 0.5, 3], [60, 1, 5], [79, 2, 8]]) {
-        identity(x, y, mu)
-      }
-      for (const [x, y, mu] of [[40, 120, 80], [55, 155, 100], [70, 190, 120]]) {
-        identity(x, y, mu)
-      }
-      for (const [x, y, mu] of [[30, 170, 140], [55, 235, 180], [80, 320, 240]]) {
-        identity(x, y, mu)
-      }
-    })
-  })
-
-  describe('.owenT()', () => {
-    it('should return reference values', () => {
-      [
-        { h: 0.0625, a: 0.25, t: 0.03891193023470137 },
-        { h: 6.5, a: 0.4375, t: 2.0005773048508314e-11 },
-        { h: 7, a: 0.96875, t: 6.399062719389869e-13 },
-        { h: 4.78125, a: 0.0625, t: 1.0632974804687463e-7 },
-        { h: 2, a: 0.5, t: 0.008625077985521507 },
-        { h: 1, a: 0.9999975, t: 0.0667418089782286 },
-        { h: 1, a: 0.5, t: 0.04306469112078537 },
-        { h: 1, a: 1, t: 0.06674188216570097 },
-        { h: 1, a: 2, t: 0.0784681869930841 },
-        { h: 1, a: 3, t: 0.0792995047488726 },
-        { h: 0.5, a: 0.5, t: 0.06448860284750375 },
-        { h: 0.5, a: 1, t: 0.1066710629614485 },
-        { h: 0.5, a: 2, t: 0.1415806036539784 },
-        { h: 0.5, a: 3, t: 0.1510840430760184 },
-        { h: 0.25, a: 0.5, t: 0.07134663382271778 },
-        { h: 0.25, a: 1, t: 0.1201285306350883 },
-        { h: 0.25, a: 2, t: 0.1666128410939293 },
-        { h: 0.25, a: 3, t: 0.1847501847929859 },
-        { h: 0.125, a: 0.5, t: 0.07317273327500386 },
-        { h: 0.125, a: 1, t: 0.1237630544953746 },
-        { h: 0.125, a: 2, t: 0.1737438887583106 },
-        { h: 0.125, a: 3, t: 0.1951190307092811 },
-        { h: 0.0078125, a: 0.5, t: 0.07378938035365545 },
-        { h: 0.0078125, a: 1, t: 0.1249951430754052 },
-        { h: 0.0078125, a: 2, t: 0.1761984774738108 },
-        { h: 0.0078125, a: 3, t: 0.1987772386442824 },
-        { h: 0.0078125, a: 10, t: 0.2340886964802671 },
-        { h: 0.0078125, a: 100, t: 0.2479460829231492 }
-      ].forEach(d => {
-        assert(equal(special.owenT(d.h, d.a), d.t))
-      })
-    })
-
-    it('should return reference values at sector boundaries with 14 significant digits', () => {
-      [
-        { h: 0.5, a: 0.025, t: 0.003510520947201143 },
-        { h: 0.5, a: 0.09, t: 0.012602627267934824 },
-        { h: 0.5, a: 0.15, t: 0.02089267814596295 },
-        { h: 0.5, a: 0.36, t: 0.04828218926612005 },
-        { h: 0.5, a: 0.5, t: 0.06448860284750375 },
-        { h: 0.5, a: 0.9, t: 0.10007270175061386 },
-        { h: 0.5, a: 0.99999, t: 0.10667044320760354 },
-        { h: 2.0, a: 0.025, t: 0.000538145641532743 },
-        { h: 2.0, a: 0.09, t: 0.0019229592641905593 },
-        { h: 2.0, a: 0.15, t: 0.0031597998016760784 },
-        { h: 2.0, a: 0.36, t: 0.006865704986743918 },
-        { h: 2.0, a: 0.5, t: 0.00862507798552151 },
-        { h: 2.0, a: 0.9, t: 0.010928598829162459 },
-        { h: 2.0, a: 0.99999, t: 0.011116267146773112 },
-        { h: 0.02, a: 0.5, t: 0.07377589505496221 },
-        { h: 0.06, a: 0.5, t: 0.07364870894373357 },
-        { h: 0.09, a: 0.5, t: 0.07347022604865686 },
-        { h: 0.125, a: 0.5, t: 0.07317273327500384 },
-        { h: 0.26, a: 0.5, t: 0.07115073619192525 },
-        { h: 0.4, a: 0.5, t: 0.06769364453308699 },
-        { h: 0.6, a: 0.5, t: 0.06077755812641674 },
-        { h: 1.6, a: 0.5, t: 0.01863684908269017 },
-        { h: 1.7, a: 0.5, t: 0.015616899411410147 },
-        { h: 2.33, a: 0.5, t: 0.004025149112463264 },
-        { h: 2.4, a: 0.5, t: 0.0033746756355782564 },
-        { h: 3.36, a: 0.5, t: 0.00018077433066393216 },
-        { h: 3.4, a: 0.5, t: 0.0001567864305170142 },
-        { h: 4.8, a: 0.5, t: 3.915533232636861e-7 }
-      ].forEach(d => {
-        assert(equal(special.owenT(d.h, d.a), d.t, 14), `owenT(${d.h}, ${d.a})`)
-      })
-    })
-
-    it('should be continuous across sector boundaries', () => {
-      const eps = 1e-9
-      const aBoundaries = [
-        { h: 0.5, a: 0.025 },
-        { h: 0.5, a: 0.09 },
-        { h: 0.5, a: 0.15 },
-        { h: 0.5, a: 0.36 },
-        { h: 0.5, a: 0.5 },
-        { h: 0.5, a: 0.9 },
-        { h: 0.5, a: 0.99999 }
-      ]
-      aBoundaries.forEach(({ h, a }) => {
-        assert(equal(special.owenT(h, a), special.owenT(h, a * (1 + eps)), 8),
-          `owenT(${h}, ${a}) a-boundary continuity`)
-      })
-      const hBoundaries = [
-        { h: 0.02, a: 0.5 },
-        { h: 0.06, a: 0.5 },
-        { h: 0.09, a: 0.5 },
-        { h: 0.125, a: 0.5 },
-        { h: 0.26, a: 0.5 },
-        { h: 0.4, a: 0.5 },
-        { h: 0.6, a: 0.5 },
-        { h: 1.6, a: 0.5 },
-        { h: 1.7, a: 0.5 },
-        { h: 2.33, a: 0.5 },
-        { h: 2.4, a: 0.5 },
-        { h: 3.36, a: 0.5 },
-        { h: 3.4, a: 0.5 },
-        { h: 4.8, a: 0.5 }
-      ]
-      hBoundaries.forEach(({ h, a }) => {
-        assert(equal(special.owenT(h, a), special.owenT(h * (1 + eps), a), 7),
-          `owenT(${h}, ${a}) h-boundary continuity`)
-      })
-    })
-  })
-
-  describe('.erf()', () => {
-    it('should return reference values', () => {
-      assert(special.erf(0) === 0)
-      ;[
-        { x: 0.5, y: 0.5204998778130465 },
-        { x: 1.0, y: 0.8427007929497149 },
-        { x: 1.5, y: 0.9661051464753108 },
-        { x: 2.0, y: 0.9953222650189527 },
-        { x: 3.0, y: 0.9999779095030014 },
-        { x: 5.0, y: 0.9999999999984626 }
-      ].forEach(d => {
-        assert(equal(special.erf(d.x), d.y), `erf(${d.x})`)
-      })
-    })
-
-    it('should satisfy erf(-x) = -erf(x)', () => {
-      [0.5, 2, 5].forEach(x => {
-        assert(equal(special.erf(-x), -special.erf(x)), `erf(-${x})`)
-      })
-    })
-  })
-
-  describe('.erfc()', () => {
-    it('should return reference values', () => {
-      assert(special.erfc(0) === 1)
-      ;[
-        { x: 0.5, y: 0.4795001221869535 },
-        { x: 1.0, y: 0.1572992070502851 },
-        { x: 2.0, y: 0.004677734981047265 },
-        { x: 3.0, y: 2.209049699858544e-5 },
-        { x: 5.0, y: 1.537459794428035e-12 },
-        { x: 10.0, y: 2.08848758376254e-45 }
-      ].forEach(d => {
-        assert(equal(special.erfc(d.x), d.y), `erfc(${d.x})`)
-      })
-    })
-
-    it('should return 0 for large positive x', () => {
-      assert(special.erfc(27) === 0)
-    })
-
-    it('should satisfy erfc(-x) = 2 - erfc(x)', () => {
-      [0.5, 2, 5].forEach(x => {
-        assert(equal(special.erfc(-x), 2 - special.erfc(x)), `erfc(-${x})`)
-      })
-    })
-
-    it('should maintain relative precision in the far tail', () => {
-      // erfc(7/sqrt(2)) appears in Normal(0,2).cdf(14); CF branch must give full precision
-      assert(equal(special.erfc(7 / Math.SQRT2), 2.559625087771669924e-12), 'erfc(7/sqrt(2))')
-    })
-  })
-
-  describe('.erfcx()', () => {
-    it('should return 1 at zero', () => {
-      assert(equal(special.erfcx(0), 1, 14))
-    })
-
-    it('should return reference values', () => {
-      ;[
-        { x: 0.5, y: 0.6156903441929259 },
-        { x: 1.0, y: 0.427583576155807 },
-        { x: 2.0, y: 0.25539567631050569 },
-        { x: 5.0, y: 0.11070463773306866 }
-      ].forEach(d => {
-        assert(equal(special.erfcx(d.x), d.y, 14), `erfcx(${d.x})`)
-      })
-    })
-
-    it('should remain finite for large x where erfc(x) underflows to 0', () => {
-      assert(special.erfc(30) === 0, 'erfc(30) underflows in float64')
-      assert(equal(special.erfcx(30), 0.018795888861416754, 14), 'erfcx(30) reference value')
-    })
-
-    it('should return reference values for negative arguments', () => {
-      ;[
-        { x: -1, y: 5.0089800807622833 },
-        { x: -3, y: 16205.988853999588 }
-      ].forEach(d => {
-        assert(equal(special.erfcx(d.x), d.y, 14), `erfcx(${d.x})`)
-      })
-    })
-  })
-
-  describe('.erfinv()', () => {
-    it('should return zero for zero argument', () => {
-      assert(special.erfinv(0) === 0)
-    })
-
-    it('should be the inverse of erf', () => {
-      [-0.9, -0.5, -0.1, 0.1, 0.5, 0.9].forEach(x => {
-        assert(equal(special.erf(special.erfinv(x)), x), `erfinv(${x})`)
-      })
-    })
-
-    it('should satisfy erfinv(-x) = -erfinv(x)', () => {
-      [0.1, 0.5, 0.9].forEach(x => {
-        assert(equal(special.erfinv(-x), -special.erfinv(x)), `erfinv(-${x})`)
-      })
-    })
-
-    it('should be accurate at small arguments where Newton iterates near zero', () => {
-      // Newton stopping criterion must handle x near 0; hybrid |dx| < EPS*max(|x|,1) is correct
-      [1e-5, 1e-8, 1e-10].forEach(x => {
-        assert(equal(special.erf(special.erfinv(x)), x), `erfinv(${x})`)
-      })
-    })
-
-    it('should be accurate at subnormal-range arguments (x^2 underflows to 0)', () => {
-      // For x ~ 1e-300, x^2 underflows to 0 so the polynomial initial guess collapses to x itself.
-      // Newton converges in one step because erf is linear at this scale; the hybrid stopping
-      // criterion must not exit prematurely before that step is taken.
-      assert(equal(special.erf(special.erfinv(1e-300)), 1e-300), 'erfinv(1e-300)')
-    })
-  })
-
-  describe('.e1()', () => {
-    it('should return boundary values', () => {
-      assert(special.e1(0) === Infinity, 'e1(0)')
-      assert(isNaN(special.e1(-1)), 'e1(-1)')
-    })
-
-    it('should return reference values in the series branch (z≤1)', () => {
-      // Reference values from mpmath at mp.dps=50
-      ;[
-        { z: 0.5, y: 0.5597735947761608 },
-        { z: 1.0, y: 0.2193839343955203 }
-      ].forEach(d => {
-        assert(equal(special.e1(d.z), d.y), `e1(${d.z})`)
-      })
-    })
-
-    it('should return reference values in the continued-fraction branch (z>1)', () => {
-      // Reference values from mpmath at mp.dps=50
-      ;[
-        { z: 2.0, y: 0.04890051070806112 },
-        { z: 5.0, y: 0.0011482955912753255 },
-        { z: 10.0, y: 4.156968929685321e-6 }
-      ].forEach(d => {
-        assert(equal(special.e1(d.z), d.y), `e1(${d.z})`)
-      })
-    })
-  })
-})
