+++ test/special/owen-t.js
@@ -1,0 +1,115 @@
+import { assert } from 'chai'
+import { describe, it } from 'mocha'
+import { equal } from '../test-utils.js'
+import * as special from '../../src/special/index.js'
+
+describe('special.owenT', () => {
+  describe('.owenT()', () => {
+    it('should return reference values', () => {
+      [
+        { h: 0.0625, a: 0.25, t: 0.03891193023470137 },
+        { h: 6.5, a: 0.4375, t: 2.0005773048508314e-11 },
+        { h: 7, a: 0.96875, t: 6.399062719389869e-13 },
+        { h: 4.78125, a: 0.0625, t: 1.0632974804687463e-7 },
+        { h: 2, a: 0.5, t: 0.008625077985521507 },
+        { h: 1, a: 0.9999975, t: 0.0667418089782286 },
+        { h: 1, a: 0.5, t: 0.04306469112078537 },
+        { h: 1, a: 1, t: 0.06674188216570097 },
+        { h: 1, a: 2, t: 0.0784681869930841 },
+        { h: 1, a: 3, t: 0.0792995047488726 },
+        { h: 0.5, a: 0.5, t: 0.06448860284750375 },
+        { h: 0.5, a: 1, t: 0.1066710629614485 },
+        { h: 0.5, a: 2, t: 0.1415806036539784 },
+        { h: 0.5, a: 3, t: 0.1510840430760184 },
+        { h: 0.25, a: 0.5, t: 0.07134663382271778 },
+        { h: 0.25, a: 1, t: 0.1201285306350883 },
+        { h: 0.25, a: 2, t: 0.1666128410939293 },
+        { h: 0.25, a: 3, t: 0.1847501847929859 },
+        { h: 0.125, a: 0.5, t: 0.07317273327500386 },
+        { h: 0.125, a: 1, t: 0.1237630544953746 },
+        { h: 0.125, a: 2, t: 0.1737438887583106 },
+        { h: 0.125, a: 3, t: 0.1951190307092811 },
+        { h: 0.0078125, a: 0.5, t: 0.07378938035365545 },
+        { h: 0.0078125, a: 1, t: 0.1249951430754052 },
+        { h: 0.0078125, a: 2, t: 0.1761984774738108 },
+        { h: 0.0078125, a: 3, t: 0.1987772386442824 },
+        { h: 0.0078125, a: 10, t: 0.2340886964802671 },
+        { h: 0.0078125, a: 100, t: 0.2479460829231492 }
+      ].forEach(d => {
+        assert(equal(special.owenT(d.h, d.a), d.t))
+      })
+    })
+
+    it('should return reference values at sector boundaries with 14 significant digits', () => {
+      [
+        { h: 0.5, a: 0.025, t: 0.003510520947201143 },
+        { h: 0.5, a: 0.09, t: 0.012602627267934824 },
+        { h: 0.5, a: 0.15, t: 0.02089267814596295 },
+        { h: 0.5, a: 0.36, t: 0.04828218926612005 },
+        { h: 0.5, a: 0.5, t: 0.06448860284750375 },
+        { h: 0.5, a: 0.9, t: 0.10007270175061386 },
+        { h: 0.5, a: 0.99999, t: 0.10667044320760354 },
+        { h: 2.0, a: 0.025, t: 0.000538145641532743 },
+        { h: 2.0, a: 0.09, t: 0.0019229592641905593 },
+        { h: 2.0, a: 0.15, t: 0.0031597998016760784 },
+        { h: 2.0, a: 0.36, t: 0.006865704986743918 },
+        { h: 2.0, a: 0.5, t: 0.00862507798552151 },
+        { h: 2.0, a: 0.9, t: 0.010928598829162459 },
+        { h: 2.0, a: 0.99999, t: 0.011116267146773112 },
+        { h: 0.02, a: 0.5, t: 0.07377589505496221 },
+        { h: 0.06, a: 0.5, t: 0.07364870894373357 },
+        { h: 0.09, a: 0.5, t: 0.07347022604865686 },
+        { h: 0.125, a: 0.5, t: 0.07317273327500384 },
+        { h: 0.26, a: 0.5, t: 0.07115073619192525 },
+        { h: 0.4, a: 0.5, t: 0.06769364453308699 },
+        { h: 0.6, a: 0.5, t: 0.06077755812641674 },
+        { h: 1.6, a: 0.5, t: 0.01863684908269017 },
+        { h: 1.7, a: 0.5, t: 0.015616899411410147 },
+        { h: 2.33, a: 0.5, t: 0.004025149112463264 },
+        { h: 2.4, a: 0.5, t: 0.0033746756355782564 },
+        { h: 3.36, a: 0.5, t: 0.00018077433066393216 },
+        { h: 3.4, a: 0.5, t: 0.0001567864305170142 },
+        { h: 4.8, a: 0.5, t: 3.915533232636861e-7 }
+      ].forEach(d => {
+        assert(equal(special.owenT(d.h, d.a), d.t, 14), `owenT(${d.h}, ${d.a})`)
+      })
+    })
+
+    it('should be continuous across sector boundaries', () => {
+      const eps = 1e-9
+      const aBoundaries = [
+        { h: 0.5, a: 0.025 },
+        { h: 0.5, a: 0.09 },
+        { h: 0.5, a: 0.15 },
+        { h: 0.5, a: 0.36 },
+        { h: 0.5, a: 0.5 },
+        { h: 0.5, a: 0.9 },
+        { h: 0.5, a: 0.99999 }
+      ]
+      aBoundaries.forEach(({ h, a }) => {
+        assert(equal(special.owenT(h, a), special.owenT(h, a * (1 + eps)), 8),
+          `owenT(${h}, ${a}) a-boundary continuity`)
+      })
+      const hBoundaries = [
+        { h: 0.02, a: 0.5 },
+        { h: 0.06, a: 0.5 },
+        { h: 0.09, a: 0.5 },
+        { h: 0.125, a: 0.5 },
+        { h: 0.26, a: 0.5 },
+        { h: 0.4, a: 0.5 },
+        { h: 0.6, a: 0.5 },
+        { h: 1.6, a: 0.5 },
+        { h: 1.7, a: 0.5 },
+        { h: 2.33, a: 0.5 },
+        { h: 2.4, a: 0.5 },
+        { h: 3.36, a: 0.5 },
+        { h: 3.4, a: 0.5 },
+        { h: 4.8, a: 0.5 }
+      ]
+      hBoundaries.forEach(({ h, a }) => {
+        assert(equal(special.owenT(h, a), special.owenT(h * (1 + eps), a), 7),
+          `owenT(${h}, ${a}) h-boundary continuity`)
+      })
+    })
+  })
+})
